### 2021-04-21 #3198 @whedon
[PRE REVIEW]: SenSARP: A Python package for pre-processing Sentinel-1 SLC data with the Sentinel-1 Toolbox
**Submitting author:** @McWhity (<a href="http://orcid.org/0000-0001-5278-8379">Thomas Weiß</a>)
**Repository:** <a href="https://github.com/multiply-org/sar-pre-processing.git" target ="_blank">https://github.com/multiply-org/sar-pre-processing.git</a>
**Version:** 0.1
