### 2021-04-23 #3198 @arfon
@McWhity – due to the small size of this code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. We should get back to you sometime next week.
