### 2021-04-28 abc5bfb
create processing list depending on orbit
