+++ sar_pre_processing/sar_pre_processor.py
@@ -17,6 +17,7 @@ from typing import List, Optional
 from .netcdf_stack import NetcdfStackCreator
 import math
 import numpy as np
+import pdb
 
 logging.getLogger().setLevel(logging.INFO)
 # Set up logging
@@ -466,53 +467,53 @@ class SARPreProcessor(PreProcessor):
 
             data_set = Dataset(file, 'r+', format="NETCDF4")
 
-            # update date information
-            try:
-                data_set.delncattr('start_date')
-                data_set.delncattr('stop_date')
-            except RuntimeError:
-                logging.warning('A runtime error has occurred')
-            data_set.setncattr_string('date', str(start_date))
+            # # update date information
+            # try:
+            #     data_set.delncattr('start_date')
+            #     data_set.delncattr('stop_date')
+            # except RuntimeError:
+            #     logging.warning('A runtime error has occurred')
+            # data_set.setncattr_string('date', str(start_date))
 
-            # extract orbit direction from metadata
-            metadata = ETree.parse(os.path.join(file_path2, filename[0:79] + '.dim'))
-            for i in metadata.findall('Dataset_Sources'):
-                for ii in i.findall('MDElem'):
-                    for iii in ii.findall('MDElem'):
-                        for iiii in iii.findall('MDATTR'):
-                            r = iiii.get('name')
-                            if r == 'PASS':
-                                orbit_dir = iiii.text
-                                data_set.setncattr_string('orbitdirection', orbit_dir)
-                            continue
+            # # extract orbit direction from metadata
+            # metadata = ETree.parse(os.path.join(file_path2, filename[0:79] + '.dim'))
+            # for i in metadata.findall('Dataset_Sources'):
+            #     for ii in i.findall('MDElem'):
+            #         for iii in ii.findall('MDElem'):
+            #             for iiii in iii.findall('MDATTR'):
+            #                 r = iiii.get('name')
+            #                 if r == 'PASS':
+            #                     orbit_dir = iiii.text
+            #                     data_set.setncattr_string('orbitdirection', orbit_dir)
+            #                 continue
 
-            # extract orbit from metadata
-            metadata = ETree.parse(os.path.join(file_path2, filename[0:79] + '.dim'))
-            for i in metadata.findall('Dataset_Sources'):
-                for ii in i.findall('MDElem'):
-                    for iii in ii.findall('MDElem'):
-                        for iiii in iii.findall('MDATTR'):
-                            r = iiii.get('name')
-                            if r == 'REL_ORBIT':
-                                relorbit = iiii.text
-                                data_set.setncattr_string('relativeorbit', relorbit)
+            # # extract orbit from metadata
+            # metadata = ETree.parse(os.path.join(file_path2, filename[0:79] + '.dim'))
+            # for i in metadata.findall('Dataset_Sources'):
+            #     for ii in i.findall('MDElem'):
+            #         for iii in ii.findall('MDElem'):
+            #             for iiii in iii.findall('MDATTR'):
+            #                 r = iiii.get('name')
+            #                 if r == 'REL_ORBIT':
+            #                     relorbit = iiii.text
+            #                     data_set.setncattr_string('relativeorbit', relorbit)
 
-            # extract frequency from metadata
-            metadata = ETree.parse(os.path.join(file_path2, filename[0:79] + '.dim'))
-            for i in metadata.findall('Dataset_Sources'):
-                for ii in i.findall('MDElem'):
-                    for iii in ii.findall('MDElem'):
-                        for iiii in iii.findall('MDATTR'):
-                            r = iiii.get('name')
-                            if r == 'radar_frequency':
-                                frequency = float(iiii.text) / 1000.
[…]
