+++ docs/notebooks/test_config_file.yaml
@@ -1,7 +1,7 @@
 ---
 gpt: /home/tweiss/snap/bin/gpt
-input_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
-output_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
+input_folder: /media/nas_data/Thomas/S1/data/Wallerfing_MNI
+output_folder: /media/nas_data/Thomas/S1/processed/MNI_2017_new_final
 region:
   lr:
     lat: 48.1
@@ -16,4 +16,4 @@ speckle_filter:
   single_filter:
     apply: 'yes'
 subset: 'yes'
-year: 2020
+year: '2017'
