### 2021-05-16 #3198 @McWhity
Right now the software is limited one satellite type (Sentinel-1). Currently two identical Sentinel-1 satellites are in orbit (S1-A since 2014 and S1-B since 2016) and provide free available microwave data of the entire world every 6 days. The satellites were designed with a lifetime of 7 years (12 years for consumables). To ensure continuous monitoring after the lifetime of S1-A and S1-B two new Sentinel-1 satellites (S1-C and S1-B) are in development and there are planned to be launched in 2022. The figure below shows the expected data volume growth by the Sentinel-1 satellites. Although the software is currently limited to one type of satellite it can be used at least until the end of this century.

![image](https://user-images.githubusercontent.com/25958669/118396892-e1b0d200-b651-11eb-9949-45bcfd1177e1.png)
Hahmann, Thomas, et al. "Sentinel-1 and Sentinel-3-OLCI PAC at DLR." Proceedings of ESA Living Planet Symposium. 2013.

The software in its current form produces a set of four different radiometric and geometric corrected sigma nought radar backscatter data.  For time series/or single time step usage:
1) multi temporal speckle filtered data
2) multi temporal speckle filtered data from different orbits normalized to one specific incidence
3) single speckle filtered data
4) single speckle filtered data normalized to one specific incidence angle

This software as a standard pre-processing scheme can be used by non expert (pre-processing) users to process Sentinel-1 data and use it for there purposes.
Expert users can easily modify the default xml-graphs (adjusting of the different processing steps for there purposes) via SNAP toolbox and use the software functions to process there data. An upload of further xml-graphs will enhance the functional scope of this software.

The “Step Forum” (https://forum.step.esa.int/search?q=pre%20processing%20Sentinel-1) shows an active community discussing how to best pre-process Sentinel-1 data.
The ReadTheDocs document of this software is used in discussions of the Step Forum (e.g. https://forum.step.esa.int/t/backscatter-normalization-to-a-pre-defined-incidence-angle-for-s-1-grd-hh-channel-data/6949/8
https://forum.step.esa.int/t/help-with-clarifying-radiometric-normalization-of-full-scene-wide-swath-incidence-angle-effects/13894)
