### 2021-05-20 #3198 @pdebuyl
@kmarkert @arthur-e would you review the article "SenSARP: A Python package for pre-processing Sentinel-1 SLC data with the Sentinel-1 Toolbox" for the Journal of Open Source Software?
