### 2021-05-20 #3198 @arthur-e
@pdebuyl Yes, I will review this paper. This would be my first review for JOSS, so let me know if there's anything else I should know beyond what's in the [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
