### 2021-05-21 #3198 @pdebuyl
Hi @arthur-e thank you for accepting to review! The page [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) is more complete than the guidelines. The [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) is also useful, but you will see it when the review starts anyway.

During the review, communication with the authors or with me goes through the review page, so that you can ask for clarifications as needed.
