### 2021-05-27 #3198 @arthur-e
@pdebuyl One thing that's not clear to me from the Reviewer Guidelines is who is responsible for creating the "[REVIEW]" Issue in `joss-reviews`. I can see the example in #1075, but it would be nice to have the markdown version of that checklist. Also, can you provide me a timeline for the review? Thanks!

The guidelines are very thorough about substance or conduct of the review, but I just can't find the logistical details about how to start one.
