### 2021-06-01 #3198 @pdebuyl
@alexamici would you review the article "SenSARP: A Python package for pre-processing Sentinel-1 SLC data with the Sentinel-1 Toolbox" for the Journal of Open Source Software?
