### 2021-06-01 #3198 @pdebuyl
Hi @alexamici thank you for getting back to me. I'll summarize the principles here:
- The journal is free to publish in and free to read (CC-BY licenses). JOSS is published by "Open Journals" and directly managed by its community.
- The papers in JOSS are short (few pages long max): the focus is on the software. For well built packages (with good documentation, tests, examples, etc) the JOSS paper is a small effort to recognize the software effort.
- Reviews take place publicly on github. You can have a look at ongoing reviews here: https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+is%3Aopen+label%3Areview
- There is a [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) which serves as a guideline.
- Reviewers are expected to install the code and run a few examples to assess its usability.

During the review, you are free to discuss with the authors and the editor via the corresponding github issue page. We have author, reviewer, and editor guidelines on our website https://joss.readthedocs.io/en/latest/index.html

I'll be happy to reply to your questions.
