### 2021-06-03 #3198 @alexamici
> Hi @alexamici I see a thumbs up on my last message, does that mean that you agree to perform the review?

The thumb-up was "OK, I'm going to study" :)

Thank you for considering me, I like the idea very much, but in the last couple of days my schedule exploded and unfortunately I'm not able to accomodate any additional commitment in the short term, not even what looks like a lightweight one. Sorry.

I will register as a reviewer anyhow and I'm happy to try and participate to reviews in the future.
