### 2021-06-04 #3198 @McWhity
I don't know the person but maybe Andreas Steinberg (https://github.com/braunfuss) would be a good candidate. He worked with the Sentinel-1 Toolbox and he has some open-source software publications (https://www.bridges.uni-kiel.de/en/team/andreas-steinberg?set_language=en)
