### 2021-06-04 #3198 @McWhity
Another possible candidate might be Pete Bunting (https://github.com/petebunting). He has also some experience in processing Sentinel-1 as well as other Earth Observation data (https://github.com/remotesensinginfo)
