### 2021-06-07 #3337 @pdebuyl
@arthur-e @braunfuss make sure to accept the invitation to the reviewers group and to have a look at the reviewer guidelines linked to at the top of this review page.

@McWhity JOSS requires that a section of the article is literally titled "statement of need", can you edit the article accordingly?

The review process will happen in this issue page, so questions to the author or to me can be added as comments here. As this is the first JOSS review for both reviewers, do not hesitate to ask questions if you have doubts about the procedure. You can also take a look at earlier reviews to get an idea of how the reviews proceed https://github.com/openjournals/joss-reviews/issues?q=is%3Aissue+label%3Aaccepted+
