### 2021-06-16 multiply-org/sar-pre-processing#12 @arthur-e
No input files found by SARPreProcessor
I am running the examples in the Jupyter Notebook at `docs/notebooks/running_test_application.ipynb` and encountering some problems.

In Step 3, the first three pre-processing steps are executed:

```py
from sar_pre_processing.sar_pre_processor import *
import warnings
warnings.filterwarnings("ignore")

processing = SARPreProcessor(config='test_config_file.yaml')
processing.create_processing_file_list()
print('start step 1')
processing.pre_process_step1()
...
```

However, no files are being found, as is evident in the output:

```
INFO:root:Found files within input folder: 0
INFO:root:Number of found files for year 2021: 0
INFO:root:Number of found files containing area of interest: 0
INFO:root:Number of found files that were double processed: 0.0
INFO:root:Number of found files with border issues: 0
INFO:root:normalisation angle not specified, default value of 35 is used for processing
INFO:root:No valid files found for pre-processing step 2.
```

It's not apparent what the `"input_folder"` parameter of the configuration file should be. I downloaded a Sentinel-1 SLC granule, single-pol (VV) IW mode image from the Alaska Satellite Facility DAAC (via NASA Earth Data) and after unzipping the archive, the directory looks like this:

```sh
$ ls -F ~/Downloads/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.SAFE
annotation/
manifest.safe
measurement/
preview/
S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.SAFE-report-20150514T155517.pdf
support/
```

Inside the `measurement` folder are three GeoTIFF files, one for each IW swath. I tried to run this with the `input_folder` set to the archive root and then again set to the `measurement` folder but in both cases no files were found.
