### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
**Regarding the input files `"input_folder`:**
- you don't need to unpack the Sentinel-1 SLC file. The code should handle the "zip-files" (unpacking, and using the required input files)
- the `input_folder` parameter should be just the path to the zip files. In your case this would probably be `~/Downloads`
