### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
**Regarding the problem that no files are being found:**

During step 3 a filter option is specified
```
`import yaml

with open('sample_config_file.yaml') as stream:
   data = yaml.safe_load(stream)

data['input_folder'] = input_folder
data['output_folder'] = output_folder
data['gpt'] = gpt_location
data['year'] = '2021'

```
In this case the year 2021 is specified as a filter option. The file you mentioned above is from 2015. Change the parameter to 2015 and it should work.
