### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
After I had a look why the download via "sentinelsat" is not working I will update the notebook to be more precise about the required input parameters
