+++ docs/notebooks/test_config_file.yaml
@@ -2,18 +2,8 @@
 gpt: /home/tweiss/snap/bin/gpt
 input_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
 output_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
-region:
-  lr:
-    lat: 48.1
-    lon: 11.9
-  ul:
-    lat: 48.4
-    lon: 11.6
 speckle_filter:
   multi_temporal:
     apply: 'yes'
     files: '5'
-  single_filter:
-    apply: 'yes'
-subset: 'yes'
-year: 2020
+
