### 2021-06-17 abb30e1
update version of sentinelsat package to 1.0.1
