### 2021-06-17 multiply-org/sar-pre-processing#12 @arthur-e
Thanks, the `year` was definitely an issue. I've fixed that, moved the compressed (ZIP) archive into a new folder and set that folder as the `input_folder`. I'm making progress as it has found the file and the CPU is doing some work when the output `start step 1` is shown.

However, it seems that Step 1 did not complete. I am getting these messages:

```
INFO:root:Found files within input folder: 1
INFO:root:Number of found files for year 2015: 1
INFO:root:area of interest not specified
INFO:root:Number of found files that were double processed: 0.0
INFO:root:Number of found files with border issues: 0
INFO:root:normalisation angle not specified, default value of 35 is used for processing
INFO:ComponentProgress:0
INFO:ComponentProgress:0
INFO:root:Process S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip with SNAP.
```

Then, after `start step 1` appears:

```
INFO:root:1
INFO:root:skip processing for /home/arthur/Downloads/Sentinel1_SLC/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip. File does not exist
INFO:root:No valid files found for pre-processing step 2.
INFO:root:skip processing for /home/arthur/Downloads/Sentinel1_SLC/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip. File /home/arthur/Downloads/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.SAFE/processed/step2/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5_GC_RC_No_Su_Co.dim does not exist.
```

While it says that `File does not exist`, the file does indeed exist (it is the compressed archive I just moved to that folder). I can confirm this with `ls` on my system. As part of Step 1, it seems to have decompressed the archive, creating a sub-folder similar to what I would see if I manually extracted the archive. However, the only content of that subfolder (so far) is the `preview` folder and the KML file for previewing the browse image. Oddly, the browse image itself was not added to this folder.

Below is the Traceback:

```
IndexError                                Traceback (most recent call last)
<ipython-input-10-dc6747795641> in <module>
     10 processing.pre_process_step2()
     11 print('start step 3')
---> 12 processing.pre_process_step3()
     13 print('start add netcdf information')
     14 processing.add_netcdf_information()

/usr/local/dev/sar-pre-processing/sar_pre_processing/sar_pre_processor.py in pre_process_step3(self)
    337
    338         # Sort file list by date (hard coded position in filename!)
--> 339         file_path, filename, file_short_name, extension = self._decompose_filename(file_list[0])
    340         file_list.sort(key=lambda x: x[len(file_path) + 18:len(file_path) + 33])
    341         file_list_old = file_list

IndexError: list index out of range
```
