### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
I know this error. This error is most likely caused by the SNAP Toolbox. There are two issues which are mentioned in `1. Requirements:`

1. The SNAP version you can download is 8.0. This version needs to be updated. Open SNAP and it will ask you if you want to install updates (Currently version 8.0.4 is the newest). Make sure that you restart SNAP because otherwise the updates won't be installed. If you use 8.0 my script will fail because a wrong internal download link is implemented in the 8.0 version of SNAP.
2. SNAP Toolbox (Linux version) need libgfortran for specific operations but currently libgfortran is not installed during the installation process of SNAP (I don't know why) therefore you might use sudo apt-get install gfortran to install the missing dependency
