### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
I will update the juypter notebook within the next hour
