### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
The software will create three processing folders ("step1", "step2", "step3)". After each step the output file should be written in the respective folder
