### 2021-06-17 c83f46d
add new version of pre_process_step3_single_file.xml to the installation process
