### 2021-06-17 fe3fd27
clean sample config file
