### 2021-06-17 multiply-org/sar-pre-processing#12 @arthur-e
Thanks, I do have `libgfortran` installed so I believe it may be related to the SNAP version instead. Would you please update the README and the Documentation to specify that SNAP version 8.0.4 (or higher?) is required? I see that [SNAP is listed as a dependency under "Module Requirements" in the README](https://github.com/multiply-org/sar-pre-processing#module-requirements) and also [in the Documentation](https://multiply-sar-pre-processing.readthedocs.io/en/master/02_Installation.html#module-requirements), but there is no version information.

It might also be helpful for users to know how to update to version 8.0.4 (I don't recall this version being available for direct download, so I assume most users will have to update their installation in the short term). For me, it was launching SNAP, and running "Help" > "Check for Updates". The download may take awhile.
