### 2021-06-17 6fd18dd
include single speckle filter files within output netcdf file
