### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
Thank you for the comment. I will update the README and Documentation accordingly.
