### 2021-06-17 multiply-org/sar-pre-processing#12 @arthur-e
Also, SNAP has to be restarted after it finishes downloading plug-ins (it will not automatically restart). JetBeans will finish the update when the application is restarted.
