### 2021-06-17 multiply-org/sar-pre-processing#12 @McWhity
I updated the jupyter notebook. The config parameters are now separated in "required" and "optional" configuration options.
Originally the code was developed to pre-process time series and not single images. But as I have seen that @arthur-e tried to use just one single image I implemented a config option to also process just one image. With only one image a multi-temporal filter is not possible therefore just a single filter will be applied during the pre-processing steps.
