### 2021-06-18 multiply-org/sar-pre-processing#12 @arthur-e
I'm still having this issue. I have:

1. Updated SNAP ("About" indicates it is now version 8.0.4);
2. `git pull` changes to this repo;
3. Checked the `libgfortran` is installed (`gfortran is already the newest version (4:9.3.0-1ubuntu2)`);
4. Reinstalled (in development mode) with `conda`;

Here are the contents of my `test_config_file.yaml` file:

```yaml
---
gpt: /usr/local/dev/esa-snap/bin/gpt
input_folder: /home/arthur/Downloads/Sentinel1_SLC/
output_folder: /home/arthur/Downloads/Sentinel1_SLC/processed
speckle_filter:
  multi_temporal:
    apply: 'no'
    files: '1'
```

And here is the output with errors; I can confirm I get the same error if I add the (correct) "year" configuration parameter to the YAML file:

```
INFO:root:Found files within input folder: 1
INFO:root:year not specified
INFO:root:area of interest not specified
INFO:root:Number of found files that were double processed: 0.0
INFO:root:Number of found files with border issues: 0
INFO:root:area of interest not specified, whole images will be processed
INFO:root:normalisation angle not specified, default value of 35 is used for processing
INFO:ComponentProgress:0
INFO:ComponentProgress:0
INFO:root:Process S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip with SNAP.

start step 1

INFO:root:1
INFO:root:skip processing for /home/arthur/Downloads/Sentinel1_SLC/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip. File does not exist
INFO:root:No valid files found for pre-processing step 2.
INFO:root:skip processing for /home/arthur/Downloads/Sentinel1_SLC/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip. File /home/arthur/Downloads/Sentinel1_SLC/processed/step2/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5_GC_RC_No_Su_Co.dim does not exist.
```

**I am noticing an error in the console (from where Jupyter Notebook was launched) that may be relevant;** starting with a kernel restart message:

```
[I 09:17:53.730 NotebookApp] Kernel restarted: 5a443a9c-75eb-493b-802d-a64353d8e4bc
[I 09:17:53.821 NotebookApp] Restoring connection for 5a443a9c-75eb-493b-802d-a64353d8e4bc:c1874a900f52437f9a1d4a35e13557ed
[I 09:17:54.830 NotebookApp] Replaying 3 buffered messages
INFO: org.esa.snap.core.gpf.operators.tooladapter.ToolAdapterIO: Initializing external tool adapters
INFO: org.esa.snap.core.util.EngineVersionCheckActivator: Please check regularly for new updates for the best SNAP experience.
Executing processing graph
INFO: org.esa.s1tbx.commons.io.ImageIOFile: Using FileCacheImageInputStream
INFO: org.hsqldb.persist.Logger: dataFileCache open start
 done.

Error: [NodeId: BandMaths(2)] Could not parse expression: 'Sigma0_VH*(sin(rad(projectedLocalIncidenceAngle))/sin(rad(incidenceAngleFromEllipsoid)))'. Undefined symbol 'Sigma0_VH'.
```

This console error appears each time the cell with `processing.create_processing_file_list()` is run.

I'll now try this for a multi-temporal dataset (executing Step 2 and downloading files first).
