### 2021-06-18 multiply-org/sar-pre-processing#13 @arthur-e
ServerError (UnsupportedOperationException) in SentinelAPI.download_all()
I'm getting a `ServerError` when trying to download files from Sentinel SciHub using the `SentinelAPI`.

I changed the date range so as to download fewer files:

```
Following products will be downloaded
['S1B_IW_SLC__1SDV_20210603T165832_20210603T165900_027195_033F94_5E37'
 'S1B_IW_SLC__1SDV_20210603T165857_20210603T165924_027195_033F94_E158'
 'S1A_IW_SLC__1SDV_20210602T170732_20210602T170759_038164_048116_EF11'
 'S1A_IW_SLC__1SDV_20210601T051818_20210601T051846_038142_048071_F417']
These 4 product need 31.2 Gb of disk space
```

And I set the download path as:

```
import os
path = os.path.join(PROJECT_ROOT, 'download')
try:
    os.makedirs(path)
except: FileExistsError
api.download_all(products, path)
```

Where `PROJECT_ROOT` is defined at the top of my notebook. I see a progress bar for the download but immediately get this error and traceback:

```
HTTPError                                 Traceback (most recent call last)
/usr/local/dev/sar-pre-processing/env/lib/python3.9/site-packages/sentinelsat/sentinel.py in _check_scihub_response(response, test_json, query_string)
   1432     try:
-> 1433         response.raise_for_status()
   1434         if test_json:

/usr/local/dev/sar-pre-processing/env/lib/python3.9/site-packages/requests/models.py in raise_for_status(self)
    942         if http_error_msg:
--> 943             raise HTTPError(http_error_msg, response=self)
    944

HTTPError: 500 Server Error: Internal Server Error for url: https://apihub.copernicus.eu/apihub/odata/v1/Products('82b56bfe-ae81-400d-9566-856ea5708c82')/$value

ServerError                               Traceback (most recent call last)
<ipython-input-6-d92535fa5ff0> in <module>
      6     os.makedirs(path)
      7 except: FileExistsError
----> 8 api.download_all(products, path)

/usr/local/dev/sar-pre-processing/env/lib/python3.9/site-packages/sentinelsat/sentinel.py in download_all(self, products, directory_path, max_attempts, checksum, n_concurrent_dl, lta_retry_delay, **kwargs)
    721             info = self.get_product_odata(pid)
    722             product_infos[pid] = info
--> 723             is_online = not self.trigger_offline_retrieval(pid)
    724             if is_online:
    725                 online_prods[pid] = info

/usr/local/dev/sar-pre-processing/env/lib/python3.9/site-packages/sentinelsat/sentinel.py in trigger_offline_retrieval(self, uuid)
    643             self.logger.error(msg)
    644             raise ServerError(msg, r)
--> 645         _check_scihub_response(r, test_json=False)
    646
    647     def download_all(

/usr/local/dev/sar-pre-processing/env/lib/python3.9/site-packages/sentinelsat/sentinel.py in _check_scihub_response(response, test_json, query_string)
   1477         elif 500 <= response.status_code < 600 or msg:
   1478             # 5xx: Server Error
-> 1479             raise ServerError(msg, response)
   1480         else:
   1481             raise SentinelAPIError(msg, response)

ServerError: HTTP status 500 Internal Server Error: UnsupportedOperationException : Error creating stream from file /mnt/s1b/dhus/ntc/incoming/x3/x3/x1/S1B_IW_SLC__1SDV_20210603T165832_20210603T165900_027195_033F94_5E37.zip
```
