### 2021-06-21 multiply-org/sar-pre-processing#13 @TonioF
The error message refers to something that is wrong server side. As you say you get a progress bar, can you please check whether the data is downloaded anyway? I reproduced your problem and found that despite the error messages everything could be downloaded.
