### 2021-06-21 d5b1e30
introduce config parameter to specify if subset to area of interest is applied
