+++ docs/notebooks/sample_config_file.yaml
@@ -16,16 +16,14 @@ gpt: /home/tweiss/snap/bin/gpt
 #pre_process_step1_border: test_pre_process_step1_border.xml
 #pre_process_step2: test_pre_process_step2.xml
 #pre_process_step3: test_pre_process_step3.xml
-
-# subset: 'yes'
-
-# region:
-#   ul:
-#     lat: 48.40
-#     lon: 11.60
-#   lr:
-#     lat: 48.10
-#     lon: 11.90
+region:
+  ul:
+    lat: 48.40
+    lon: 11.60
+  lr:
+    lat: 48.10
+    lon: 11.90
+subset: 'no'
 
 speckle_filter:
   multi_temporal:
