### 2021-06-21 multiply-org/sar-pre-processing#13 @arthur-e
Is the download asynchronous, then? Because I look at the Jupyter Kernel status and it is idle (i.e., the `ServerError` seems to be issue right before the program exits).

According to [a previous report about this error message](https://stackoverflow.com/questions/67582137/http-status-500-internal-server-error-unsupportedoperationexception-error-cr), the issue is intermittent, but I haven't yet found a time where it succeeds.
