### 2021-06-21 324332b
further implementation of single image processing
