+++ sar_pre_processing/sar_pre_processor.py
@@ -286,6 +286,14 @@ class SARPreProcessor(PreProcessor):
             logging.info('No valid files found for pre-processing step 2.')
             return
 
+        if len(file_list) == 1:
+            logging.info('Single image, no co-register of images necessary')
+            self.config.single_file = 'yes'
+            return
+
+        if len(file_list) > 1:
+            self.config.single_file = 'no'
+
         # Set Master image for co-registration
         master = file_list[0]
 
@@ -324,31 +332,32 @@ class SARPreProcessor(PreProcessor):
         if not os.path.exists(self.config.output_folder_step3):
             os.makedirs(self.config.output_folder_step3)
 
-        # Create filelist with all to be processed images
-        if self.file_list is None:
-            logging.info('no file list specified, therefore all images in output folder step2 will be processed')
-            file_list = self._create_file_list(self.config.output_folder_step2, '*.dim')
-        else:
-            file_list = []
-            for list in self.file_list:
-                for file in list:
-                    file_path, filename, file_short_name, extension = self._decompose_filename(file)
-                    new_file_name = os.path.join(self.config.output_folder_step2, file_short_name +
-                                                 self.name_addition_step1 + self.name_addition_step2 + '.dim')
-                    if os.path.exists(new_file_name) is True:
-                        file_list.append(new_file_name)
-                    else:
-                        logging.info(f'skip processing for {file}. File {new_file_name} does not exist.')
-
-        # Sort file list by date (hard coded position in filename!)
-        file_path, filename, file_short_name, extension = self._decompose_filename(file_list[0])
-        file_list.sort(key=lambda x: x[len(file_path) + 18:len(file_path) + 33])
-        file_list_old = file_list
-
-        if self.config.speckle_filter.multi_temporal.apply == 'yes':
+        if (self.config.speckle_filter.multi_temporal.apply == 'yes') and (self.config.single_file == 'no'):
             # Check if XML file for pre-processing step 3 is specified
             assert self.config.pre_process_step3 is not None, \
                 'ERROR: path of XML file for pre-processing step 3 is not not specified'
+
+            # Create filelist with all to be processed images
+            if self.file_list is None:
+                logging.info('no file list specified, therefore all images in output folder step2 will be processed')
+                file_list = self._create_file_list(self.config.output_folder_step2, '*.dim')
+            else:
+                file_list = []
+                for list in self.file_list:
+                    for file in list:
+                        file_path, filename, file_short_name, extension = self._decompose_filename(file)
+                        new_file_name = os.path.join(self.config.output_folder_step2, file_short_name +
+                                                     self.name_addition_step1 + self.name_addition_step2 + '.dim')
+                        if os.path.exists(new_file_name) is True:
+                            file_list.append(new_file_name)
+                        else:
+                            logging.info(f'skip processing for {file}. File {new_file_name} does not exist.')
+
+            # Sort file list by date (hard coded position in filename!)
+            file_path, filename, file_short_name, extension = self._decompose_filename(file_list[0])
+            file_list.sort(key=lambda x: x[len(file_path) + 18:len(file_path) + 33])
+            file_list_old = file_list
+
 
             # loop to apply multi-temporal filtering
             # vv and vh polarisation are separated
@@ -439,10 +448,34 @@ class SARPreProcessor(PreProcessor):
                 logging.info(datetime.now())
 
 
-        elif self.config.speckle_filter.multi_temporal.apply == 'no':
+        elif self.config.single_file == 'yes':
+            if self.config.speckle_filter.multi_temporal.apply == 'yes':
[…]
