### 2021-06-21 f1c5ce2
put region and subset configuration parameter together
