+++ docs/notebooks/sample_config_file.yaml
@@ -17,13 +17,13 @@ gpt: /home/tweiss/snap/bin/gpt
 #pre_process_step2: test_pre_process_step2.xml
 #pre_process_step3: test_pre_process_step3.xml
 region:
+  subset: 'no'
   ul:
     lat: 48.40
     lon: 11.60
   lr:
     lat: 48.10
     lon: 11.90
-subset: 'no'
 
 speckle_filter:
   multi_temporal:
