+++ sar_pre_processing/file_list_sar_pre_processing.py
@@ -329,7 +329,7 @@ class SARList(object):
 
         # list with all zip files that contain area of interest
         try:
-            if self.config.region.subset == 'yes'
+            if self.config.region.subset == 'yes':
                 lower_right_y = self.config.region['lr']['lat']
                 upper_left_y = self.config.region['ul']['lat']
                 upper_left_x = self.config.region['ul']['lon']
