### 2021-06-21 7907a48
add single_file parameter to sample configuration file
