### 2021-06-21 babcd6f
fix naming of incidence angle for single file processing
