### 2021-06-21 94fc06c
added option for applying single speckle filter and no multi temporal speckle filter to a time series
