### 2021-06-21 93f6bd2
if processing of single file is detected. config parameter for applying multi temporal speckle filtering is changed to "no"
