### 2021-06-21 multiply-org/sar-pre-processing#12 @McWhity
I was not able to reproduce this error. However, as pre-processing step 2 (co-registration of images) is not needed in the case of processing one single image, I rewrote the new implemented feature of just processing one single image by skipping pre-processing step 2.
New feature:
- skipping pre-processing step 2 in case of single image processing
- software can is no able to detect if just one image is given as input data. If single image is detected incorrect configuration parameters like
```
single_file: 'no'
speckle_filter:
  multi_temporal:
    apply: 'yes'
```
will be overwritten and following statements will be given
```
INFO:root:Single image, no co-register of images necessary
INFO:root:multi temporal filter cannot applied to a single image, just single speckle filter is applied
```

The new implementation should fix the problem.
Please try it out and let me know. Thank you
