### 2021-06-21 #3337 @pdebuyl
Hi @arthur-e thank you very much for the review. With JOSS, we do not expect a "minor/major revision" assessment. Providing a detailed report, as you have done, participates in the constructive review process which will take place in this issue.
