### 2021-06-21 #3337 @arthur-e
@pdebuyl Thanks! [The "Review criteria"](https://joss.readthedocs.io/en/latest/review_criteria.html#what-happens-if-the-software-i-m-reviewing-doesn-t-meet-the-joss-criteria) may need to be updated as it still mentions "Minor" and "Major" revisions.
