### 2021-06-22 multiply-org/sar-pre-processing#13 @TonioF
No, the datasets are downloaded successively. To be clear, I get the same error - even multiple times - but in the end the progress bar is moving and it is downloaded. You get the error after the progress bar appears?
