### 2021-06-22 multiply-org/sar-pre-processing#13 @arthur-e
Yes, I see the error message after the progress bar appears.
