### 2021-06-22 multiply-org/sar-pre-processing#13 @arthur-e
To be more specific, the progress bar reads:

```
Fetching archival status: 0% [--------Progress bar-----------] 0/4 [00:00<?, ? products/s]
```

Where `[--Progress bar--]` stands in for the graphical display of an (empty) progress bar.
