### 2021-06-22 multiply-org/sar-pre-processing#12 @arthur-e
I'm getting a different error now--making progress! I'm getting an `IndexError` on [Line 546 in `sar_pre_processor.py`](https://github.com/multiply-org/sar-pre-processing/blob/ceb29523572cb43046756cac4ec26faabe9e9066/sar_pre_processing/sar_pre_processor.py#L546):

```py
file_path, filename, file_short_name, extension = self._decompose_filename(file_list[0])
```

It appears that `file_list` is an empty list.

Above this line, there is a conditional that sets up `file_list` as an empty list, [starting on Line 530](https://github.com/multiply-org/sar-pre-processing/blob/ceb29523572cb43046756cac4ec26faabe9e9066/sar_pre_processing/sar_pre_processor.py#L530):

```py
if self.file_list is None:
    logging.info('no file list specified, therefore all images in output folder step2 will be processed')
    file_list = self._create_file_list(self.config.output_folder_step2, '*.dim')
else:
    file_list = []
    ...
```

For me, `self.file_list` is a tuple with two elements:

```py
(['/home/arthur/Downloads/Sentinel1_SLC/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5.zip'],
[])
```

**I'm not sure if that second empty list is another potential issue.** Because `self.file_list` is not `None`, `file_list` is initialized as an empty list. On the first pass iterating through `self.file_list` there is one filename passed to `self._decompose_filename()`:

```py
# Starting on Line 537
file_path, filename, file_short_name, extension = self._decompose_filename(file)
new_file_name = os.path.join(self.config.output_folder_step2, file_short_name +
                             self.name_addition_step1 + self.name_addition_step2 + '.dim')
if os.path.exists(new_file_name) is True:
    file_list.append(new_file_name)
else:
    logging.info(f'skip processing for {file}. File {new_file_name} does not exist.')
```

**So, `new_file_name` gets created as:**

```
'/home/arthur/Downloads/Sentinel1_SLC/step2/S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5_GC_RC_No_Su_Co.dim'
```

**The `step2` folder was created at some point but the folder is empty;** the file `S1A_IW_SLC__1SSV_20150514T134812_20150514T134839_005918_0079FD_90D5_GC_RC_No_Su_Co.dim` isn't created so the check `os.path.exists(new_file_name)` will always fail. Then, `file_list` does not get `new_file_name` appended to it, so it remains empty.
