### 2021-06-25 multiply-org/sar-pre-processing#13 @TonioF
This error could come from problems with the sentinelsats versions. Which version of sentinelsat are qou using?

Also, can you confirm that the 'products' that are the result of the api.query()-call is an OrderedDict?
