### 2021-06-30 multiply-org/sar-pre-processing#13 @arthur-e
I'm using version 1.0.1 of `sentinelsat`. It was installed automatically as part of the `sar-pre-processing` installation. And, yes, the result of the `api.query()` call, `products` is an instance of `collections.OrderedDict`.
