### 2021-07-07 multiply-org/sar-pre-processing#13 @TonioF
I'm still struggling to reproduce this error. My best guesses are to try again or use this bit instead of api.download_all(products, path):

for product in products.keys():
    api.download(product, path)

From what I could gather this is a problem that some users experience with sentinelsat / copernicus hub , and there are times when the request could be executed later. Anyway, could you try with the snippet above?
