### 2021-07-09 2b7a468
fixing documentation error within usage part "ModuleNotFoundError"
