+++ environment.yml
@@ -15,6 +15,7 @@ dependencies:
   - pytest>=4.6 # testing
   - gdal>=3.0.4
   - jupyter
+  - matplotlib
   - pip
   - pip:
     - sentinelsat==1.0.1
