### 2021-07-09 83f1316
fixing documentation error within usage part "ModuleNotFoundError"
