+++ docs/requirements.txt
@@ -11,9 +11,3 @@ nbsphinx
 ipykernel
 sentinelsat
 matplotlib
-
-python:
-  version: 3.6
-  install:
-    - method: sar_pre_processing
-      path: .
