+++ .readthedocs.yml
@@ -13,3 +13,6 @@ python:
   version: 3.6
   install:
     - requirements: docs/requirements.txt
+    - method: setuptools
+      path: .
+
