### 2021-07-09 3fac00d
include missing references.bib file
