+++ docs/references.bib
@@ -1,0 +1,104 @@
+@techreport{SentinelsPODteam,
+    title = {Sentinels POD Service File Format Specifications},
+    author = {Sentinels POD team},
+    institution = {ESA},
+    address = {Ref: GMES-GSEG-EOPG-FS-10-0075},
+    year = {2016},
+}
+@techreport{Miranda,
+    title = {Radiometric Calibration of S-1 Level-1 Products Generated by the S-1 IPF},
+    author = {Miranda, N. and Meadows, P.J.},
+    institution = {ESA},
+    address = {Ref: ESA-EOPG-CSCOP-TN-0002},
+    year = {2015},
+}
+@techreport{Sentinel1Team,
+    title = {Sentinel-1 User Handbook},
+    author = {Sentinel-1 Team},
+    institution = {ESA},
+    address = {Ref: GMES-S1OP-EOPG-TN-13-0001},
+    year = {2013},
+    type = {Manual}
+}
+@techreport{Sentinel1ProductDefinition,
+    title = {Sentinel-1 Product Definition},
+    author = {Bourbigot, H. and Johnsen, H. and Piantanida R.},
+    institution = {ESA},
+    address = {Ref: S1-RS-MDA-53-7440},
+    year = {2015},
+}
+@article{d2007burst,
+    title={Burst-mode SARs for wide-swath surveys},
+    author={d'Aria, D and Zan, F De and Giudici, D and Guarnieri, A Monti and Rocca, F},
+    journal={Canadian Journal of Remote Sensing},
+    volume={33},
+    number={1},
+    pages={27--38},
+    year={2007},
+    publisher={NRC Research Press Ottawa, Canada}
+}
+@article{de2006topsar,
+    title={TOPSAR: Terrain observation by progressive scans},
+    author={De Zan, Francesco and Guarnieri, AM Monti},
+    journal={IEEE Transactions on Geoscience and Remote Sensing},
+    volume={44},
+    number={9},
+    pages={2352--2360},
+    year={2006},
+    publisher={IEEE}
+}
+@manual{small2008,
+    title = {Guide to ASAR Geocoding},
+    author = {Small, D. and Schubert, A.},
+    organization = {University of Zurich},
+    address = {Ref: RSL-ASAR-GC-AD},
+    year = {2008}
+}
+@ARTICLE{kellndorfer,
+    author={J. M. Kellndorfer and L. E. Pierce and M. C. Dobson and F. T. Ulaby},
+    journal={IEEE Transactions on Geoscience and Remote Sensing},
+    title={Toward consistent regional-to-global-scale vegetation characterization using orbital SAR systems},
+    year={1998},
+    volume={36},
+    number={5},
+    pages={1396-1411},
+    doi={10.1109/36.718844},
+    ISSN={0196-2892},
+    month={Sep}
+}
+@book{ulaby1982microwave,
+  title={Microwave remote sensing: active and passive. Vol. 2, Radar remote sensing and surface scattering and emission theory},
+  author={Ulaby, Fawwaz T and Moore, Richard K and Fung, Adrian K},
+  year={1982},
+  publisher={Addison-Wesley Reading, MA}
+}
+@book{woodhouse2005introduction,
+  title={Introduction to microwave remote sensing},
+  author={Woodhouse, Iain H},
+  year={2005},
+  publisher={CRC press}
+}
+@ARTICLE{ardila,
+author={J. P. Ardila and V. Tolpekin and W. Bijker},
+journal={IEEE Geoscience and Remote Sensing Letters},
+title={Angular Backscatter Variation in L-Band ALOS ScanSAR Images of Tropical Forest Areas},
+year={2010},
+volume={7},
+number={4},
+pages={821-825},
[…]
