### 2021-07-09 9c34f78
add bibtex file within conf of RtD
