### 2021-08-17 c8abe26
include information about updating SNAP from 8.0 to 8.0.4
