### 2021-08-19 18a2812
update installation process
