+++ setup.py
@@ -23,6 +23,8 @@ def read(*names, **kwargs):
     ) as fh:
         return fh.read()
 
+with open('docs/requirements.txt') as ff:
+    required = ff.read().splitlines()
 
 setup(name='multiply-sar-pre-processing',
       version=__version__,
@@ -33,6 +35,7 @@ setup(name='multiply-sar-pre-processing',
       author_email='weiss.thomas@lmu.de',
       url='https://github.com/multiply-org/sar-pre-processing',
       packages=['sar_pre_processing', 'sar_pre_processing.default_graphs'],
+      install_requires=required,
       package_data={'sar_pre_processing.default_graphs': ['pre_process_step1.xml', 'pre_process_step1_border.xml',
                                                           'pre_process_step2.xml', 'pre_process_step3.xml', 'pre_process_step3_single_file.xml']
     },
