### 2021-08-21 347cc04
include option to run user defined xml graphs
