### 2021-08-23 b4c3cb8
rewrite stagging of netcdf
