### 2021-08-23 0005d57
update sentinelsat version
