+++ environment.yml
@@ -18,4 +18,4 @@ dependencies:
   - matplotlib
   - pip
   - pip:
-    - sentinelsat==1.0.1
+    - sentinelsat==1.1.0
