+++ docs/03_Usage.rst
@@ -1,11 +1,289 @@
 Usage
-=========
-The follwing juypter notebook shows how to use this python package.
-You can also `view the notebook on nbviewer`__.
+======
 
-__ https://nbviewer.jupyter.org/github/multiply-org/sar-pre-processing/tree/master/docs/notebooks/
+Requirements
+--------------
 
-.. toctree::
-    :maxdepth: 2
+- Installation of SenSARP
+- Installation of ESA's SNAP Toolbox version >8.0.3
+    - Currently only SNAP version 8.0 can be downloaded from the `ESA website <https://step.esa.int/main/download/snap-download/>`_. To update SNAP to a version >8.0.3 please start the SNAP software. You will be asked if you want to search for update. After the updates are installed you need to restart SNAP to initialize the installed updates.
+    - SNAP Toolbox need libgfortran for specific operations but currently libgfortran is not installed during the installation process of SNAP therefore you might use::
 
-    notebooks/running_test_application.ipynb
+        sudo apt-get install gfortran
+
+- Sentinel-1 SLC data
+    - Instruction how to download Sentinel 1 data are given in :ref:`Download_S1_data`.
+
+.. _Download_S1_data:
+
+Download Sentinel-1 data
+--------------------------
+
+Option 1: Download data from Sentinel Data Hub via python package sentinelsat
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Create Account (`<https://scihub.copernicus.eu/dhus/#/self-registration>`_) and change user and password below::
+
+    # connect to the API
+    from sentinelsat import SentinelAPI, read_geojson, geojson_to_wkt
+    from datetime import date
+    user = 'username'
+    password = 'password'
+    # initialize settings
+    api = SentinelAPI(user, password)
+
+Search for available data::
+
+    # search by polygon (MNI test site coordinates), time, and SciHub query keywords
+    footprint = geojson_to_wkt(read_geojson('coordinates_mni.geojson'))
+    products = api.query(footprint,
+                         date=('20210601', '20210615'),
+                         platformname='Sentinel-1',
+                         producttype='SLC')
+    print('Following products will be downloaded')
+    print(api.to_dataframe(products).title.values)
+
+    print('These {} product need {} Gb of disk space'.format(len(products), api.get_products_size(products)))
+
+Start download process (Attention: might take a while and data will requries some free disk space)::
+
+    # download all results from the search
+    # files will be downloaded to specified path
+    import os
+    path = "/path/to/data/"
+    try:
+        os.makedirs(path)
+    except: FileExistsError
+    api.download_all(products, path)
+
+
+Option 2: Download data from NASA Earth Data Search
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+You can search for Sentinel-1 data `here <https://search.earthdata.nasa.gov/search>`_.
+Instructions how to search and download data from NASA Earth Data can be found `here <https://earthdata.nasa.gov/faq/earthdata-search-faq>`_.
+
+Examples
+----------
+
+Apply default pre-processing chain of SenSARP to a time-series
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Set paths for
+
+- input_folder (path to stored Sentinel-1 SLC data (zip files) e.g. “~/Downloads”)
+- output_folder (path where processed data will be stored e.g. “~/output”)
+- gpt_loction (gpt is located in the bin folder of your SNAP installation)::
+
+    import os
+    input_folder = os.path.expanduser(path)
+    output_folder = os.path.expanduser(path)
[…]
