+++ docs/04_Theory.rst
@@ -1,9 +1,9 @@
 .. _ProcessingChain:
 
-SNAP Pre-Processing Chain
-==========================
+Default Pre-Processing Chain
+============================
 
-Main goal: Geometric and radiometric corrected Sigma nought backscatter values for usage within radiative transfer models like Water Cloud Model.
+The default pre-processing chain was developed to process Sentinel-1 SLC data to provide geometric and radiometric corrected Sigma nought backscatter values. Among other things the processed data can be used within different radiative transfer models (e.g. Integration Equation Model :cite:`fung_backscattering_1992`, Oh Model :cite:`oh_empirical_1992`, :cite:`yisok_oh_quantitative_2004`, Dubois Model :cite:`dubois_measuring_1995`, Water Cloud Model :cite:`attema_vegetation_1978`, Single Scattering Radiative Transfer :cite:`ulaby_microwave_2014`, :cite:`weis_evaluation_2020`) to retrieve different land surface and/or vegetation parameters.
 
 Overview
 --------
@@ -29,7 +29,7 @@ Sentinel-1 Level-1 SLC data
 The preprocessing work-flow of :numref:`work_flow_step_1` is based on Sentinel-1 Level-1 SLC data. Among some other sources Sentinel-1 data can be downloaded from ESA's Copernicus Open Access Hub (`<https://scihub.copernicus.eu/>`_).
 
 
-Sentinel-1 Level-1 SLC data are generated by the operational ESA Instrument Processing Facility (IPF). The SLC products are situated in slant range geometry. The slant range geometry is the natural radar one and is defined by the line-of-sight distance of the radar system to each reflecting object. The SLC product consists of focused SAR data in zero-Doppler orientation. Furthermore, for geo-referencing orbit and attitude information directly provided by the satellite are stored within the SLC product. Moreover the SAR data is corrected for errors caused by the well known azimuth bi-static delay, elevation antenna pattern and range spreading loss :cite:`Sentinel1Team`. In contrary to Level-1 Ground Range Detected (GRD) products SLC data preserve the real and imaginary part of the backscatter signal and contain therefore also the phase information :cite:`Sentinel1Team`. The IPF is generating SLC data for all available acquisition modes (StripMap (SM), Interferometric Wide (IW), Extra Wide (EW), and Wave (WV)) of the Sentinel-1 satellites. Further information about Sentinel-1 Level-1 products are gathered in ESA's Sentinel-1 User Handbook :cite:`Sentinel1Team` available at `<https://earth.esa.int/documents/247904/685163/Sentinel-1_User_Handbook>`_.
[…]
