+++ docs/requirements.txt
@@ -6,4 +6,4 @@ sphinxcontrib-bibtex>=0.4 # docs
 pytest>=4.6 # testing
 jupyter
 matplotlib
-sentinelsat==1.0.1
+sentinelsat==1.1.0

[…]
