+++ docs/download.rst
@@ -1,0 +1,78 @@
+2. Download sample data from Sentinel Data Hub
+----------------------------------------------
+
+Option 1: Download data from Sentinel Data Hub via python package sentinelsat
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Create Account (https://scihub.copernicus.eu/dhus/#/self-registration)
+and change user and password below.
+
+.. code:: ipython3
+
+    # connect to the API
+    from sentinelsat import SentinelAPI, read_geojson, geojson_to_wkt
+    from datetime import date
+    user = 'user'
+    password = 'password'
+    # initialize settings
+    api = SentinelAPI(user, password)
+
+Search for available data
+
+.. code:: ipython3
+
+    # search by polygon (MNI test site coordinates), time, and SciHub query keywords
+    footprint = geojson_to_wkt(read_geojson('coordinates_mni.geojson'))
+    products = api.query(footprint,
+                         date=('20210601', '20210602'),
+                         platformname='Sentinel-1',
+                         producttype='SLC')
+    print('Following products will be downloaded')
+    print(api.to_dataframe(products).title.values)
+    
+    print('These {} product need {} Gb of disk space'.format(len(products), api.get_products_size(products)))
+
+**Code output:**
+
+.. code:: ipython3
+
+    Following products will be downloaded
+    ['S1A_IW_SLC__1SDV_20210601T051818_20210601T051846_038142_048071_F417']
+    These 1 product need 7.97 Gb of disk space
+
+
+Start download process (**Attention: might take a while and data will
+requries some free disk space)**
+
+.. code:: ipython3
+
+    # download all results from the search
+    # files will be downloaded to specified path
+    import os
+    path = os.path.expanduser('~/Desktop/data')
+    try:
+        os.makedirs(path)
+    except: FileExistsError
+    api.download_all(products, path)
+
+
+**Code output:**
+
+.. code:: ipython3
+
+    Downloading products:   100%|########| 1/1 [49:04<00:00, 2964.23s/product]
+    Downloading S1A_IW_SLC__1SDV_20210601T051818_20210601T051846_038142_048071_F417.zip:   100%|########| 4.6/4.6…
+
[…]
