+++ docs/requirements.rst
@@ -1,0 +1,21 @@
+1. Requirements
+------------------
+
+-  Installation of SenSARP
+-  Installation of ESA’s SNAP Toolbox version >8.0.3
+
+   -  Currently only SNAP version 8.0 can be downloaded from the ESA
+      website (https://step.esa.int/main/download/snap-download/). To
+      update SNAP to a version >8.0.3 please start the SNAP software.
+      You will be asked if you want to search for update. After the
+      updates are installed you need to restart SNAP to initialize the
+      installed updates.
+   -  SNAP Toolbox need libgfortran for specific operations but
+      currently libgfortran is not installed during the installation
+      process of SNAP therefore you might use
+      ``sudo apt-get install gfortran``
+
+-  Sentinel-1 SLC data
+
+   -  Instruction how to download Sentinel 1 data are given in Section 2
+
