+++ docs/examples.rst
@@ -1,0 +1,682 @@
+3. Examples
+------------
+Example 1: Use default processing graph to pre-process single SAR image
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Set paths for
+    - input_folder (path to stored Sentinel-1 SLC data (zip files) e.g. “~/Downloads”)
+    - output_folder (path where processed data will be stored e.g. “~/output”)
+    - gpt_loction (gpt is located in the bin folder of your SNAP installation)’
+
+.. code:: ipython3
+
+    input_folder = path
+    output_folder = path
+    gpt_location = os.path.expanduser('~/snap/bin/gpt')
+
+Create config file with information about input folder, output folder
+and gpt path
+
+.. code:: ipython3
+
+    import yaml
+    
+    with open('sample_config_file.yaml') as stream:
+       data = yaml.safe_load(stream)
+    
+    data['input_folder'] = input_folder
+    data['output_folder'] = output_folder
+    data['gpt'] = gpt_location
+    
+    with open('test_config_file.yaml', 'wb') as stream:
+       yaml.safe_dump(data, stream, default_flow_style=False, 
+                      explicit_start=True, allow_unicode=True, encoding='utf-8')
+
+Optional config options for subsetting
+
+.. code:: ipython3
+
+    with open('test_config_file.yaml') as stream:
+       data = yaml.safe_load(stream)
+    
+    ## Define region of interest
+    data['region']['lr']['lat'] = 48.2 # lower right latitude
+    data['region']['lr']['lon'] = 11.9 # lower right longitude
+    data['region']['ul']['lat'] = 48.4 # upper left latitude
+    data['region']['ul']['lon'] = 11.6 # upper left longitude
+    data['region']['subset'] = 'yes'
+    
+    with open('test_config_file.yaml', 'wb') as stream:
+       yaml.safe_dump(data, stream, default_flow_style=False, 
+                      explicit_start=True, allow_unicode=True, encoding='utf-8')
+
+Start pre-processing steps
+
+.. code:: ipython3
+
+    from sar_pre_processing.sar_pre_processor import *
+    import warnings
+    warnings.filterwarnings("ignore")
+    
+    processing = SARPreProcessor(config='test_config_file.yaml')
+    processing.create_processing_file_list()
+    print('start step 1')
+    processing.pre_process_step1()
+    print('start step 2')
+    processing.pre_process_step2()
+    print('start step 3')
+    processing.pre_process_step3()
+    print('start add netcdf information')
+    processing.add_netcdf_information()
+    print('start create netcdf stack')
+    processing.create_netcdf_stack()
+
+
+**Code output:**
+
+.. code:: ipython3
+
+    INFO:root:Found files within input folder: 1
+    INFO:root:year not specified
+    INFO:root:area of interest not specified
+    INFO:root:Number of found files that were double processed: 0.0
+    INFO:root:Number of found files with border issues: 0
+    INFO:root:area of interest specified
+    INFO:root:normalisation angle not specified, default value of 35 is used for processing
+    INFO:ComponentProgress:0
+    INFO:ComponentProgress:0
[…]
