+++ sar_pre_processing/netcdf_stack.py
@@ -70,8 +70,6 @@ class NetcdfStackCreator(object):
         self.longitude = self.dataset.createVariable('lon', np.float32, ('lon'))
 
         # create 3-D variables (localIncidenceangle ... )
-        # self.localIncidenceAngle = self.dataset.createVariable('localIncidenceAngle', np.float32,('time','lat','lon'), fill_value=-99999)
-        # self.localIncidenceAngle.units = 'degree'
 
         for i in data.variables.keys():
             if i == 'lat':
@@ -81,36 +79,12 @@ class NetcdfStackCreator(object):
             elif i == 'crs':
                 pass
             else:
-                if user_defined_graphs == 'yes':
+                if self.user_defined_graphs == 'yes':
                     self.dataset.createVariable(i[:-15], np.float32,('time','lat','lon'), fill_value=-99999)
                     #self.dataset[i[:-15]].units = 'linear'
                 else:
                     self.dataset.createVariable(i, np.float32,('time','lat','lon'), fill_value=-99999)
                     self.dataset[i].units = 'linear'
-
-
-
-        # if self.temporal_filter == 'yes':
-
-        #     self.sigma0_vv_multi = self.dataset.createVariable('sigma0_vv_multi', np.float32,('time','lat','lon'), fill_value=-99999)
-        #     self.sigma0_vv_multi.units = 'linear'
-        #     self.sigma0_vh_multi = self.dataset.createVariable('sigma0_vh_multi', np.float32,('time','lat','lon'), fill_value=-99999)
-        #     self.sigma0_vh_multi.units = 'linear'
-
-        #     self.sigma0_vv_norm_multi = self.dataset.createVariable('sigma0_vv_norm_multi', np.float32,('time','lat','lon'), fill_value=-99999)
-        #     self.sigma0_vv_norm_multi.units = 'linear'
-        #     self.sigma0_vh_norm_multi = self.dataset.createVariable('sigma0_vh_norm_multi', np.float32,('time','lat','lon'), fill_value=-99999)
-        #     self.sigma0_vh_norm_multi.units = 'linear'
-
-        # self.sigma0_vv_single = self.dataset.createVariable('sigma0_vv_single', np.float32,('time','lat','lon'), fill_value=-99999)
-        # self.sigma0_vv_single.units = 'linear'
-        # self.sigma0_vh_single = self.dataset.createVariable('sigma0_vh_single', np.float32,('time','lat','lon'), fill_value=-99999)
-        # self.sigma0_vh_single.units = 'linear'
-
-        # self.sigma0_vv_norm_single = self.dataset.createVariable('sigma0_vv_norm_single', np.float32,('time','lat','lon'), fill_value=-99999)
-        # self.sigma0_vv_norm_single.units = 'linear'
-        # self.sigma0_vh_norm_single = self.dataset.createVariable('sigma0_vh_norm_single', np.float32,('time','lat','lon'), fill_value=-99999)
-        # self.sigma0_vh_norm_single.units = 'linear'
 
     def stacking(self):
         """stack all files in one netcdf file"""
@@ -162,23 +136,9 @@ class NetcdfStackCreator(object):
                 elif i == 'crs':
                     pass
                 else:
-                    if user_defined_graphs == 'yes':
+                    if self.user_defined_graphs == 'yes':
                         self.dataset[i[:-15]][index,:,:] = data.variables[i][:]
                     else:
                         self.dataset[i][index,:,:] = data.variables[i][:]
-
-
-            # self.localIncidenceAngle[index,:,:] = data.variables['theta'][:]
-            # if self.temporal_filter == 'yes':
[…]
