+++ docs/03_Usage.rst
@@ -7,6 +7,8 @@ Usage
     requirements
     download
     examples
+    visualization
+    config
 
 
 
