+++ docs/examples.rst
@@ -680,3 +680,343 @@ Start pre-processing steps
 
     Scene 7 of 7
     /home/test/Desktop/data/step3/S1A_IW_SLC__1SDV_20210609T165916_20210609T165943_038266_0483FE_CD3F_GC_RC_No_Su_Co_speckle.nc
+
+
+Example 3: Use user defined xml graphs to process Sentinel-1 data
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Set paths for
+    - input_folder (path to stored Sentinel-1 SLC data (zip files) e.g. “~/Downloads”)
+    - output_folder (path where processed data will be stored e.g. “~/output”)
+    - gpt_loction (gpt is located in the bin folder of your SNAP installation)’
+
+.. code:: ipython3
+
+    input_folder = path
+    output_folder = path
+    gpt_location = os.path.expanduser('~/snap/bin/gpt')
+
+Create config file with information about input, output and gpt location
+
+.. code:: ipython3
+
+    import yaml
+
+    with open('sample_config_file.yaml') as stream:
+       data = yaml.safe_load(stream)
+
+    data['input_folder'] = input_folder
+    data['output_folder'] = output_folder
+    data['gpt'] = gpt_location
+
+    with open('test_config_file.yaml', 'wb') as stream:
+       yaml.safe_dump(data, stream, default_flow_style=False,
+                      explicit_start=True, allow_unicode=True, encoding='utf-8')
+
+Expert user might create their own processing chain and use the
+functionally of SenSARP to automate the processing on scale
+1. Create a user defined xml graph via SNAP’s Graph Builder
+    - information how to use SNAP’s Graph Builder can be found at the step forum (https://forum.step.esa.int/t/graph-builder/5403)
+2. Modifiy user defined xml graph to be used by SenSARP - Needed modification
+    - Placeholder for input image within xml graph need to be specified
+    - Placeholder for output image within xml graph need to be specified
+    - If subset operator is used within user defined xml graph a placeholder for subset extent need to be specified
+
+Functionallity of SenSARP that can be used:
+    - Filter option (year, area of interest)
+    - default pre_process_step2 (co-registration of time series images)
+    - creation of netCDF4 stack containing all processed images and information about orbit direction, relative orbit, satellite name
+
+.. code:: ipython3
+
+    # Load xml graph
+    import xml.etree.ElementTree as ETree
+    user_xml_graph = '/home/test/Desktop/sar-pre-processing/sar_pre_processing/user_defined_graphs/expert_user.xml'
+    tree = ETree.parse(user_xml_graph)
+    root = tree.getroot()
+
+    # find Operators Read, Write and Subset
+    read = root.find("./node/[operator='Read']")
+    read2 = read.find("./parameters[@class='com.bc.ceres.binding.dom.XppDomElement']")
+    write = root.find("./node/[operator='Write']")
+    subset = root.find("./node/[operator='Subset']")
+
+.. code:: ipython3
+
+    # Old Version of user defined xml graph
+    print(ETree.tostring(read, encoding='utf8').decode('utf8'))
+    print(ETree.tostring(write, encoding='utf8').decode('utf8'))
+    print(ETree.tostring(subset, encoding='utf8').decode('utf8'))
+
+
+**Output code:**
+
[…]
