+++ docs/default_process_single_image.rst
@@ -1,0 +1,406 @@
+
+1. Requirements
+---------------
+
+-  Installation of SenSARP
+-  Installation of ESA’s SNAP Toolbox version >8.0.3
+
+   -  Currently only SNAP version 8.0 can be downloaded from the ESA
+      website (https://step.esa.int/main/download/snap-download/). To
+      update SNAP to a version >8.0.3 please start the SNAP software.
+      You will be asked if you want to search for update. After the
+      updates are installed you need to restart SNAP to initialize the
+      installed updates.
+   -  SNAP Toolbox need libgfortran for specific operations but
+      currently libgfortran is not installed during the installation
+      process of SNAP therefore you might use
+      ``sudo apt-get install gfortran``
+
+-  Sentinel-1 SLC data
+
+   -  Instruction how to download Sentinel 1 data are given in Section 2
+
+2. Download sample data from Sentinel Data Hub
+----------------------------------------------
+
+Option 1: Download data from Sentinel Data Hub via python package sentinelsat
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Create Account (https://scihub.copernicus.eu/dhus/#/self-registration)
+and change user and password below.
+
+.. code:: ipython3
+
+    # connect to the API
+    from sentinelsat import SentinelAPI, read_geojson, geojson_to_wkt
+    from datetime import date
+    user = 'mcwhity'
+    password = 'dS1dfSSDHasap'
+    # initialize settings
+    api = SentinelAPI(user, password)
+
+Search for available data
+
+.. code:: ipython3
+
+    # search by polygon (MNI test site coordinates), time, and SciHub query keywords
+    footprint = geojson_to_wkt(read_geojson('coordinates_mni.geojson'))
+    products = api.query(footprint,
+                         date=('20210601', '20210602'),
+                         platformname='Sentinel-1',
+                         producttype='SLC')
+    print('Following products will be downloaded')
+    print(api.to_dataframe(products).title.values)
+    
+    print('These {} product need {} Gb of disk space'.format(len(products), api.get_products_size(products)))
+
+Output:
+    Following products will be downloaded
+    ['S1A_IW_SLC__1SDV_20210601T051818_20210601T051846_038142_048071_F417']
+    These 1 product need 7.97 Gb of disk space
+
+
+Start download process (**Attention: might take a while and data will
+requries some free disk space)**
+
+.. code:: ipython3
+
+    # download all results from the search
+    # files will be downloaded to specified path
+    import os
+    path = os.path.expanduser('~/Desktop/data')
+    try:
+        os.makedirs(path)
+    except: FileExistsError
+    api.download_all(products, path)
+
+
+Output:
+    Downloading products:   0%|          | 0/1 [00:00<?, ?product/s]
+    Downloading S1A_IW_SLC__1SDV_20210601T051818_20210601T051846_038142_048071_F417.zip:   0%|          | 0.00/4.6…
+    MD5 checksumming:   0%|          | 0.00/4.64G [00:00<?, ?B/s]
+
[…]
