+++ docs/03_Usage.rst
@@ -4,11 +4,11 @@ Usage
 .. toctree::
     :maxdepth: 2
 
-    requirements
-    download
-    examples
-    visualization
-    config
+    031_requirements
+    032_download
+    033_examples
+    034_visualization
+    035_config
 
 
 
