+++ docs/requirements.txt
@@ -4,11 +4,11 @@ numpy==1.18.1
 pyproj==2.6.0
 pytest==4.6
 PyYAML==5.4
-setuptools==49.2.1
+setuptools
 m2r
 sphinxcontrib-bibtex
 nbsphinx
 ipykernel
-sentinelsat==1.1.0
+sentinelsat>=1.1.0
 matplotlib
 jupyter

[…]
