### 2021-08-29 fd2636b
try inserting ipynb to RtD
