+++ docs/03_Usage.rst
@@ -9,6 +9,7 @@ Usage
     033_examples
     034_visualization
     035_config
+    default_process_time_series.ipynb
 
 
 
