+++ docs/default_process_time_series.rst
@@ -1,1125 +1,0 @@
-Example of running sar-pre-processing package
-=============================================
-
-1. Requirements
----------------
-
--  Installation of SenSARP
--  Installation of ESA’s SNAP Toolbox version >8.0.3
-
-   -  Currently only SNAP version 8.0 can be downloaded from the ESA
-      website (https://step.esa.int/main/download/snap-download/). To
-      update SNAP to a version >8.0.3 please start the SNAP software.
-      You will be asked if you want to search for update. After the
-      updates are installed you need to restart SNAP to initialize the
-      installed updates.
-   -  SNAP Toolbox need libgfortran for specific operations but
-      currently libgfortran is not installed during the installation
-      process of SNAP therefore you might use
-      ``sudo apt-get install gfortran``
-
--  Sentinel-1 SLC data
-
-   -  Instruction how to download Sentinel 1 data are given in Section 2
-
-2. Download sample data from Sentinel Data Hub
-----------------------------------------------
-
-Option 1: Download data from Sentinel Data Hub via python package sentinelsat
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-Create Account (https://scihub.copernicus.eu/dhus/#/self-registration)
-and change user and password below.
-
-.. code:: ipython3
-
-    # connect to the API
-    from sentinelsat import SentinelAPI, read_geojson, geojson_to_wkt
-    from datetime import date
-    user = 'mcwhity'
-    password = 'dS1dfSSDHasap'
-    # initialize settings
-    api = SentinelAPI(user, password)
-
-Search for available data
-
-.. code:: ipython3
-
-    # search by polygon (MNI test site coordinates), time, and SciHub query keywords
-    footprint = geojson_to_wkt(read_geojson('coordinates_mni.geojson'))
-    products = api.query(footprint,
-                         date=('20210601', '20210610'),
-                         platformname='Sentinel-1',
-                         producttype='SLC')
-    print('Following products will be downloaded')
-    print(api.to_dataframe(products).title.values)
-    
-    print('These {} product need {} Gb of disk space'.format(len(products), api.get_products_size(products)))
-
-
-.. parsed-literal::
-
-    Following products will be downloaded
-    ['S1A_IW_SLC__1SDV_20210609T165916_20210609T165943_038266_0483FE_CD3F'
-     'S1A_IW_SLC__1SDV_20210609T165941_20210609T170008_038266_0483FE_08A0'
-     'S1B_IW_SLC__1SDV_20210607T051737_20210607T051804_027246_034125_2C2A'
-     'S1A_IW_SLC__1SDV_20210606T052628_20210606T052655_038215_04828D_89C3'
-     'S1B_IW_SLC__1SDV_20210603T165832_20210603T165900_027195_033F94_5E37'
-     'S1B_IW_SLC__1SDV_20210603T165857_20210603T165924_027195_033F94_E158'
-     'S1A_IW_SLC__1SDV_20210602T170732_20210602T170759_038164_048116_EF11'
[…]
