+++ docs/notebooks/test_config_file.yaml
@@ -1,9 +1,19 @@
 ---
-gpt: /home/tweiss/snap/bin/gpt
-input_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
-output_folder: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/docs/notebooks/test_files
+gpt: /home/wodan/snap/bin/gpt
+input_folder: /path/to/data/
+output_folder: /path/to/data/
+region:
+  lr:
+    lat: 48.2
+    lon: 11.9
+  subset: 'yes'
+  ul:
+    lat: 48.4
+    lon: 11.6
+single_file: 'no'
 speckle_filter:
   multi_temporal:
     apply: 'yes'
     files: '5'
-
+use_user_defined_graphs: 'yes'
+year: '2021'
