### 2021-08-29 5f8373e
update class description, use newer version of sentinelsat package, include expert user xml file
