+++ docs/031_config.rst
@@ -1,5 +1,5 @@
-5. Configuration file
-----------------------
+Explanation of configuration file
+----------------------------------
 
 .. code-block:: yaml
 
@@ -8,21 +8,21 @@
     ## Necessary parameters
     #-----------------------
     ### Input folder with SAR data (zip format)
-    input_folder: '/media/tweiss/Daten/test_mysentinelapi' # values: string
+    input_folder: '/media/test/Desktop/data' # values: string
 
     ### Output folder to store the pre-processed data
-    output_folder: '/media/tweiss/Daten/test_mysentinelapi' # values: string
+    output_folder: '/media/test/Desktop/data' # values: string
 
     ### Location of SNAP's graph-processing-tool
     gpt: /home/tweiss/snap/bin/gpt
 
     ## Optional parameters
     #-----------------------
-    ### Year of interest
+    ### Year of interest (only images of the specified year will be processed)
     #~~~~~~~~~~~~~~~~~~~~
     year: 2021 # values: integer
 
-    ### Area of interest
+    ### Area of interest (only images containing the specified year will be processed)
     #~~~~~~~~~~~~~~~~~~~
     region:
       subset: 'yes' # values: 'no' or 'yes'
@@ -44,20 +44,20 @@
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     normalisation_angle: 35 # values: float
 
-    ### Single file processing
+    ### Single file processing (if filelist contains only one file single_file option is automatically set to 'yes')
     #~~~~~~~~~~~~~~~~~~~~~~~~~
     single_file: 'yes' # values: 'no' or 'yes'
 
     ### Usage of user defined xml graphs
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-    #### Specification is user defined xml graph should be used
+    #### Specification if user defined xml graph should be used
     use_user_definde_graphs: 'no' # values: 'no' or 'yes'
 
     #### Location of user defined xml files for processing
-    xml_graph_path: /media/tweiss/Work/GIT/GitHub/multiply-org/sar-pre-processing/sar_pre_processing/user_defined_graphs
+    xml_graph_path: /media/test/Desktop/sar-pre-processing/sar_pre_processing/user_defined_graphs
 
     ##### file names of user defined xml graphs
-    pre_process_step1: expert_user.xml
+    pre_process_step1: example_of_expert_user.xml
 
 
 
