+++ docs/conf.py
@@ -203,7 +203,7 @@ html_static_path = ['_static']
 # html_file_suffix = None
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'multiply_sar_pre_processing_doc'
+htmlhelp_basename = 'SenSARP'
 
 
 # -- Options for LaTeX output --------------------------------------------------
