### 2021-08-29 multiply-org/sar-pre-processing#13 @McWhity
There is now a new version of sentinelsat available (1.1.0). This should hopefully fix the problem. I included in the [jupyter notebook](https://github.com/multiply-org/sar-pre-processing/blob/master/docs/notebooks/default_process_single_image.ipynb) some more information how to download Sentinel-1 data. You can download Sentinel-1 data from the Copernicus Data Hub manually or via the new version of the sentinelsat package. If the Copernicus Data Hub is not working an alternative is provided with the Alaska Satellite Facility.
