### 2021-08-30 #3337 @McWhity
> I have completed my review.
>
> I am unsure whether the submission requires "Minor" or "Major" revisions. I would tentatively propose that "Major revisions" are required only because I could not fully evaluate the functionality of the software due to an outstanding technical issue.
> ## Comments to Authors
>
> Overall, SenSARP seems to provide a valuable, very high-level processing chain for Sentinel 1 SLC data. However, there are two main areas in need of improvement: 1) Key functionality issue(s) need to be resolved; and, 2) The documentation should be improved and expanded to provide more details to users that may want to deviate from the prescribed pre-processing chain.
>
> **First, there is [an unresolved issue](https://github.com/multiply-org/sar-pre-processing/issues/12) as to the functionality of pre-processing Step 1.** While unresolved, this issue prevents the software from functioning and, therefore, impedes further evaluation of its functionality.
>
> I'd like to see the Documentation updated to provide end-users with more information on how to customize their processing chain. In addition to adding more details on the end-user functionality, there are a number syntax/ typographical issues with the documentation (e.g., missing citations in the ReadTheDocs compilation) that should fixed. Also, while SenSARP seems to have automated tests (and CI through Travis CI), the authors should add instructions to the README on how users can run tests themselves. I was able to run the two Python scripts in the `tests/` directory without error, but I didn't get any feedback in the console so I can't be sure. I've written more about proposed Documentation improvements in the section on Documentation, below.
>
Thank you for pointing this out.
- Syntax/typographical issues (missing citations, ReadTheDocs compilation) within documentation are resolved
- Documentation is updated to provide more end-user information (see answers to review points on the documentation)
- Information how to run test within [Readme](https://github.com/multiply-org/sar-pre-processing/blob/master/README.md) and [ReadtheDocs documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/01_Introduction.html) were added

> It is also difficult to determine how much the SenSARP Toolbox improves upon the ESA SNAP Sentinel-1 Toolbox. For instance, [`SARPreProcessor.pre_process_step1()`](https://multiply-sar-pre-processing.readthedocs.io/en/latest/_modules/sar_pre_processing/sar_pre_processor.html#SARPreProcessor.pre_process_step1) consists of reading a configuration file, setting defaults, and using `subprocess.call()` to activate the corresponding SNAP Sentinel-1 Toolbox functions. The authors should revise the accompanying paper as well as the documentation to more explicitly state how this software improves upon the SNAP toolbox. If the major improvement is that `SenSARP` offers a programmatic API to the SNAP toolbox functions, that would be a meaningful contribution, but it should be made clear. While the authors do make the case that "an automatic or manual pre-processing of Sentinel-1 images is needed" (Lines 19-20), it would be good to know for sure that the SNAP toolbox _does not_ offer such an automated pipeline, so that the value of `SenSARP` is obvious. **In short, what are the relative advantages of SenSARP over the SNAP Toolbox, which is a dependency?** I believe answering this question will better address the journal requirement: "A Statement of Need section that clearly illustrates the research purpose of the software." The journal recommends a section titled "Statement of Need" in the paper, though I think a section header is not necessary as long as the need is clearly described towards the beginning of the paper.
>
> I think that SenSARP could be characterized differently in the JOSS paper and in the software's documentation/ README to make it clear that it provides a specific pre-processing chain with sensible(?) defaults--I would not call it a "toolbox" myself, [because it has very few individual functions/ classes for end-users](https://multiply-sar-pre-processing.readthedocs.io/en/latest/annex_technical.html), but I do not dispute the use of that term. I think it might be more useful (and accurate?) to characterize `SenSARP` as a "push-button" or "turn-key" solution (or some other analogy) to automated pre-processing of Sentinel-1 SLC data. It seems that rather than a multi-functional programming API, `SenSARP` makes it easy to complete a complex but fairly rigid processing chain to multiple images. My first impression of `SenSARP` was that it would have multiple end-points/ functions/ classes for representing different kinds of data and processing steps; but, [it seems to consist primarily of three end-user functions](https://multiply-sar-pre-processing.readthedocs.io/en/latest/annex_technical.html), `pre_process_step1()`, `pre_process_step2()`, `pre_process_step3()`. So, I think of `SenSARP` more as a pipeline rather than a "toolbox."
>
Thank you for these very helpful remarks. A section "Statement of Need" addressing the mentioned shortcomings is now added to the [Paper](https://github.com/multiply-org/sar-pre-processing/blob/master/docs/paper.pdf), the [Readme](https://github.com/multiply-org/sar-pre-processing/blob/master/README.md), and the [Documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/01_Introduction.html). SenSARP is not characterized as "toolbox" anymore but as suggested SenSARP is now called a "push-button" / "pipeline" solution

> Finally, the authors should address the journal requirement on Community Guidelines: "There should be clear guidelines for third-parties wishing to: Contribute to the software; Report issues or problems with the software; Seek support." Some repositories include a `CONTRIBUTING` (Markdown or RST) file with detailed instructions about how Pull Requests and Issues should be formatted. Alternatively, this information could be added to a section of the `README`.

Thank you for pointing this out. A section "Support, contributing and testing" was added to the [Readme](https://github.com/multiply-org/sar-pre-processing/blob/master/README.md) and the [documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/01_Introduction.html)

> ### Functionality
>
> **As mentioned above, there is [an unresolved issue](https://github.com/multiply-org/sar-pre-processing/issues/12) as to the functionality of pre-processing Step 1.**
>
Please see answer to [an unresolved issue](https://github.com/multiply-org/sar-pre-processing/issues/12)

> [Another outstanding issue](https://github.com/multiply-org/sar-pre-processing/issues/13) that is unrelated may be a problem with the ESA data service and not with SenSARP. However, this issue (with the ESA data service) seems to severely impact part of the SenSARP functionality, in that data must be downloaded manually from somewhere else. The authors should address this shortcoming and provide some guidance to end-users in the Documentation and the Jupyter Notebook. **Is there somewhere else users can get Sentinel 1 SLC data?** I used NASA Earth Data Search.
>
Thank you for this remark. You are right SenSARP can only be used if the Sentinel-1 data is available. Two options are now introduced within the [jupyter notebook](https://github.com/multiply-org/sar-pre-processing/blob/master/docs/notebooks/default_process_single_image.ipynb).
- Option 1: Downloading data manually or via sentinelsat from Copernicus data hub.
- Option 2: Downloading data from the Alaska Satellite Facility (ASF uses the same credentials als NASA Earth Data Search).
The two options are also introduced within the [documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/notebooks/default_process_single_image.html)

> **Regarding Installation...**
>
>     * **One key change that is needed: Version details for the SNAP Toolbox dependency.** In the review, [it became apparent that Version 8.0.4 (or greater?) is required](https://github.com/multiply-org/sar-pre-processing/issues/12#issuecomment-863323367) but in the installation instructions there is no such information--and an unsupported version will be installed by default when the SNAP Toolbox is downloaded. Users should be told explicitly (how) to upgrade to 8.0.4.
>
Thank you for this comment. Information that a update is needed as well as information how to update SNAP is given within the [Readme](https://github.com/multiply-org/sar-pre-processing/blob/master/README.md), the beginning of the [jupyter notebook](https://github.com/multiply-org/sar-pre-processing/blob/master/docs/notebooks/default_process_single_image.ipynb) and the installation part of the [documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/02_Installation.html#further-information)

>     * I would prefer it if there were some way to install the software without using `conda`, which can be a pain to work with (and very slow). In this case, `conda` took several minutes to install the repository (spent a lot of time on "Solving environment"). Ideally, the `setup.py` script would list the package dependencies so that the repository (and its dependencies) could be installed with either `distutils` or `pip`. This can still be done within a `conda` virtual environment or, better yet, a `virtualenv` environment.
>
>     * It would be helpful if the installation instructions could describe/ link to `conda` installation instructions, both for Anaconda and `miniconda`.
>
>     * A few more details on how to download and install the ESA SNAP Sentinel toolbox would be helpful to users.

Thank you for this remark.
- Instruction how to install SenSARP via virtualenv and python were added. As the installation of GDAL can be very tricky, GDAL needs to be installed separately. Instruction how to do this is were added
- Information how to download and install Anaconda and Miniconda were added
- More details on how to download and install the ESA SNAP Sentinel toolbox were added

Please see installation part of [readme](https://github.com/multiply-org/sar-pre-processing/blob/master/README.md) and [documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/02_Installation.html#installation-with-virtualenv-and-python)

>
>
> ### Documentation
>
> [In the "Usage" section,](https://multiply-sar-pre-processing.readthedocs.io/en/latest/notebooks/running_test_application.html#2.-Download-sample-data-from-Sentinel-Data-Hub) there seem to be some errors that occurred at the time the documentation was compiled (e.g., `ModuleNotFoundError: No module named 'sentinelsat'`). Fixing these and showing the correct output should help users to understand what the correct behavior of the software looks like.
>
Thank you for pointing this out. The ["Usage"](https://multiply-sar-pre-processing.readthedocs.io/en/latest/03_Usage.html) section was revised. Now showing three jupyter notebook examples (default processing of a single image, default processing a time series, processing a time series with a user defined xml graph) and an explanation of the config file.

> I really like the detailed description of each processing step, however, the "Practical implementation" sections seem incomplete. While most users might use the defaults for `pre_process_step1()`, `pre_process_step2()`, and `pre_process_step3()`, is it possible that a user might want to skip or change a step? For instance, maybe a user wants to skip radiometric correct in this pipeline and do it differently later. Backscatter normalization is also "optional" [but it's not clear from the documentation](https://multiply-sar-pre-processing.readthedocs.io/en/latest/04_Theory.hhttps://multiply-sar-pre-processing.readthedocs.io/en/latest/04_Theory.html#backscatter-normalisation-optionaltml#backscatter-normalisation-optional) how to turn it on/ off. I would like to see these details fully described in the "Practical implementation" sections, perhaps with a minimal example (of how to configure the YAML file for that aim?).
>
Thank you for pointing this out. As the theory of default pre-processing steps are explained within "Practical implementation" the section was renamed to "Default processing chain". For non-expert users skipping of a step is not intended right now. I understand the confusion about the "optional" behind the "backscatter normalization" step. The backscatter normalization as well as the multi-temporal filter are applied in any case. However, the final netcdf file contains layers of
- single speckle filtered radiometric and geometric corrected sigma nought backscatter
- multi speckle filtered radiometric and geometric corrected sigma nought backscatter
- single speckle filtered radiometric and geometric corrected sigma nought backscatter normalized to a specific incidence angle
- multi speckle filtered radiometric and geometric corrected sigma nought backscatter normalized to a specific incidence angle

The missing information about which layers and which content is provided by the final netcdf file a section explaining the [output layers](https://multiply-sar-pre-processing.readthedocs.io/en/master/04_Theory.html#output-layers-of-final-netcdf-stack-file) was added. Another section ["Folder and data creation during pre-processing steps"](https://multiply-sar-pre-processing.readthedocs.io/en/master/04_Theory.html#folder-and-data-creation-during-pre-processing-steps) explaining which folders are created during the default pre-processing was added as well.
Regarding the skipping of specific processing steps. A new [jupyter notebook](https://github.com/multiply-org/sar-pre-processing/blob/master/docs/notebooks/use_user_defined_graphs.ipynb) explaining how expert users can modify the default xml-processing graphs or create their own processing graphs and use the SenSARP functionalities was added to the ["Usage"](https://multiply-sar-pre-processing.readthedocs.io/en/master/notebooks/use_user_defined_graphs.html#3.-Use-user-defined-xml-graph-to-process-Sentinel-1-data-(test-case-with-Sentinel-1-GRD-product)) section.

> There are also some minor syntax issues with the documentation that could be fixed, e.g., [the Theory/purpose section of the section on TOPSAR Deburst](https://multiply-sar-pre-processing.readthedocs.io/en/latest/04_Theory.html#id12) reads: "More detailed information are provided in [], [] and []." I think that _none_ of the in-line citations are appearing on ReadTheDocs for any section.
>
Thank you for pointing this out. Syntax issues were fixed.

> Finally, I think the README could be updated to provide explicit instructions to users for running the tests, including information as to how to evaluate they ran correctly--for instance, the users should see _no output_ in the console if the tests succeeded, correct?

Thank you for pointing this out. As mentioned previously a section "Support, contributing and testing" was added. If you run "pytest" in the terminal you should receive an output if the test did run correctly or if they failed.

> ### Software Paper
>
> As discussed above, I believe the authors could do more to distinguish SenSARP from the SNAP Toolbox. Specifically:
>
>     * How does SenSARP improve uponthe SNAP Toolbox? It would appear that SenSARP provides [the pre-processing functionality of the SNAP Toolbox](http://step.esa.int/main/toolboxes/sentinel-1-toolbox/s1tbx-features/) via `subprocess` calls to SNAP. This may be a substantial improvement over the graphical user interface (GUI) of SNAP through the ability to automate these steps at scale. If this is the primary contribution of SenSARP, the authors should explicitly state so in the paper.
>
Thank you for your very helpful remarks. As already mentioned a section "Statement of need" clarifying the mentioned shortcomings was added. Another feedback in this matter is highly appreciated.

>     * Is SenSARP a fairly rigid pre-processing pipeline or is it a multi-functionality API for creating customized GPT pipelines? Neither the documentation nor the Jupyter Notebook provide examples of how the pipeline might be customized to skip or change certain pre-processing steps.

Thank you for your very helpful remarks. A new [juypter notebook](https://multiply-sar-pre-processing.readthedocs.io/en/latest/notebooks/use_user_defined_graphs.html#3.-Use-user-defined-xml-graph-to-process-Sentinel-1-data-(test-case-with-Sentinel-1-GRD-product)) explaining how expert users can customize or use their own pre-processing chain was added.
