### 2021-10-06 #3337 @pdebuyl
Hi @danclewley thank you for accepting to review. The general guidelines for reviewing are here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

In a nutshell:
- Read our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
- Take a look at the package, install it, check and run the tests and examples.
- Go over your checklist (in the first message of this page) and use it as a guide to perform the review.

Don't forget to accept the invitation to the group of reviewers, you need to accept it to be able to check the items in the checklist.
