### 2021-10-12 3fef84d
run NDVI processing
