### 2021-10-29 #3337 @pdebuyl
Hi @danclewley thank you for getting on to this.

Regarding scholarly efforts, you can find the criteria here: https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort

During the pre-review phase, during which the question was raised by Daniel Katz, the author already provided extra information https://github.com/openjournals/joss-reviews/issues/3198#issuecomment-841809933

Among the criteria for scholary effort, the paper mentions existing applications of the package, which also count toward the evalution. I let the author reply for further motivation if necessary.
