### 2021-11-05 #3337 @danclewley
Apologies this has taken so long.

I'm happy to recommend **accept with minor revisions** if the paper is updated to clarify who it is aimed at and how it differs from other packages.
