### 2021-11-05 #3337 @pdebuyl
Hi @danclewley thank for the feedback.

To proceed, I would like to ask you to fill the checklist for all items that you agree on. This is part of the JOSS review process and it in general also a nice way for the authors to track the progress of the review. To accept the paper, once you are ok, I would also like an explicit statement of acceptance here. (This is not part of JOSS guidelines but without this, with the back and forth messages and the absence of notifications for checklist editing, I find it harder to conclude the review).
