+++ docs/references.bib
@@ -11,7 +11,7 @@
 	month = jul,
 	year = {1995},
 	note = {Conference Name: IEEE Transactions on Geoscience and Remote Sensing},
-	keywords = {1.5 to 11 GHz, airborne radar, Content based retrieval, dual frequency method, empirical algorithm, geophysical techniques, hydrological techniques, hydrology, imaging radar, Information retrieval, measurement technique, moisture measurement, Moisture measurement, radar applications, Radar imaging, Radar measurements, Radar remote sensing, remote sensing, remote sensing by radar, scatterometer, soil, Soil measurements, soil moisture, Soil moisture, spaceborne radar, Spaceborne radar, UHF SHF microwave, Vegetation mapping},
+	keywords = {Soil moisture, hydrological techniques, moisture measurement, remote sensing by radar, soil, Soil measurements, spaceborne radar, Spaceborne radar, Vegetation mapping, geophysical techniques, radar applications, remote sensing, soil moisture, Radar measurements, Radar remote sensing, scatterometer, hydrology, airborne radar, Content based retrieval, Moisture measurement, 1.5 to 11 GHz, dual frequency method, empirical algorithm, imaging radar, Information retrieval, measurement technique, Radar imaging, UHF SHF microwave},
 	pages = {915--926},
 	file = {Submitted Version:/home/wodan/snap/zotero-snap/common/Zotero/storage/ENPA2M8L/Dubois et al. - 1995 - Measuring soil moisture with imaging radars.pdf:application/pdf;IEEE Xplore Abstract Record:/home/wodan/snap/zotero-snap/common/Zotero/storage/2PGVAXYX/406677.html:text/html},
 }
@@ -28,7 +28,7 @@
 	month = mar,
 	year = {2004},
 	note = {Conference Name: IEEE Transactions on Geoscience and Remote Sensing},
-	keywords = {airborne radar, airborne synthetic aperture radar, atmospheric humidity, Backscatter, bare soil surfaces, Content based retrieval, Equations, image retrieval, inverse problems, Moisture measurement, multipolarized radar observations, Parameter estimation, Radar polarimetry, remote sensing by radar, Rough surfaces, semiempirical polarimetric backscattering model, Soil measurements, Soil moisture, soil moisture content retrieval, surface height, surface roughness, Surface roughness, synthetic aperture radar, topography (Earth), volumetric soil moisture content},
+	keywords = {Soil moisture, remote sensing by radar, Soil measurements, synthetic aperture radar, Backscatter, surface roughness, atmospheric humidity, Radar polarimetry, airborne radar, airborne synthetic aperture radar, bare soil surfaces, Content based retrieval, Equations, image retrieval, inverse problems, Moisture measurement, multipolarized radar observations, Parameter estimation, Rough surfaces, semiempirical polarimetric backscattering model, soil moisture content retrieval, surface height, Surface roughness, topography (Earth), volumetric soil moisture content},
 	pages = {596--601},
 	file = {IEEE Xplore Abstract Record:/home/wodan/snap/zotero-snap/common/Zotero/storage/7GEYUQLP/1273591.html:text/html},
 }
@@ -59,7 +59,7 @@ Introduction; Electromagnetic Wave Propagation; Remote-Sensing Antennas; Microwa
 	month = mar,
 	year = {1978},
 	note = {Conference Name: Radio Science},
-	keywords = {Agriculture, Clouds, Mathematical model, Radar imaging, Soil, Vegetation},
+	keywords = {Vegetation, Mathematical model, Clouds, Agriculture, Radar imaging, Soil},
 	pages = {357--364},
 	file = {IEEE Xplore Abstract Record:/home/wodan/snap/zotero-snap/common/Zotero/storage/NU2W645J/7767291.html:text/html},
 }
[…]
