+++ docs/paper.bib
@@ -1,81 +1,343 @@
 
-@misc{noauthor_multiply_nodate,
-  title = {Multiply | {{Horizon}} 2020 Project},
-  file = {/home/wodan/snap/zotero-snap/common/Zotero/storage/U2GGWNYA/multiply.obs-website.eu-de.otc.t-systems.com.html},
-  language = {en-GB}
+@article{dubois_measuring_1995,
+	title = {Measuring soil moisture with imaging radars},
+	volume = {33},
+	issn = {1558-0644},
+	doi = {10.1109/36.406677},
+	abstract = {An empirical algorithm for the retrieval of soil moisture content and surface root mean square (RMS) height from remotely sensed radar data was developed using scatterometer data. The algorithm is optimized for bare surfaces and requires two copolarized channels at a frequency between 1.5 and 11 GHz. It gives best results for kh/spl les/2.5, /spl mu//sub /spl upsi///spl les/35\%, and /spl theta//spl ges/30/spl deg/. Omitting the usually weaker hv-polarized returns makes the algorithm less sensitive to system cross-talk and system noise, simplifies the calibration process and adds robustness to the algorithm in the presence of vegetation. However, inversion results indicate that significant amounts of vegetation (NDVI{\textgreater}0.4) cause the algorithm to underestimate soil moisture and overestimate RMS height. A simple criteria based on the /spl sigma//sub hv//sup 0///spl sigma//sub vv//sup 0/ ratio is developed to select the areas where the inversion is not impaired by the vegetation. The inversion accuracy is assessed on the original scatterometer data sets but also on several SAR data sets by comparing the derived soil moisture values with in-situ measurements collected over a variety of scenes between 1991 and 1994. Both spaceborne (SIR-C) and airborne (AIRSAR) data are used in the test. Over this large sample of conditions, the RMS error in the soil moisture estimate is found to be less than 4.2\% soil moisture.{\textless}{\textgreater}},
+	number = {4},
+	journal = {IEEE Transactions on Geoscience and Remote Sensing},
+	author = {Dubois, P.C. and van Zyl, J. and Engman, T.},
+	month = jul,
+	year = {1995},
+	note = {Conference Name: IEEE Transactions on Geoscience and Remote Sensing},
+	keywords = {Soil moisture, hydrological techniques, moisture measurement, remote sensing by radar, soil, Soil measurements, spaceborne radar, Spaceborne radar, Vegetation mapping, geophysical techniques, radar applications, remote sensing, soil moisture, Radar measurements, Radar remote sensing, scatterometer, hydrology, airborne radar, Content based retrieval, Moisture measurement, 1.5 to 11 GHz, dual frequency method, empirical algorithm, imaging radar, Information retrieval, measurement technique, Radar imaging, UHF SHF microwave},
+	pages = {915--926},
+	file = {Submitted Version:/home/wodan/snap/zotero-snap/common/Zotero/storage/ENPA2M8L/Dubois et al. - 1995 - Measuring soil moisture with imaging radars.pdf:application/pdf;IEEE Xplore Abstract Record:/home/wodan/snap/zotero-snap/common/Zotero/storage/2PGVAXYX/406677.html:text/html},
 }
 
-@misc{noauthor_multiscale_nodate,
-  title = {{{MULTIscale SENTINEL}} Land Surface Information Retrieval {{PLatform}} | {{MULTIPLY Project}} | {{H2020}}},
[…]
