+++ docs/paper.bib
@@ -80,7 +80,7 @@
 
 @inproceedings{werner_gamma_2000,
     address = {Gothenburg Sweden},
-    title = {{GAMMA} {SAR} {AND} {INTERFEROMETRIC} {PROCESSING} {SOFTWARE}},
+    title = {{GAMMA} {SAR} {and} {Interferometric} {Processing} {Software}},
     abstract = {The GAMMA Modular SAR Processor (MSP), Interferometric SAR Processor (ISP), Differential Interferometry and Geocoding Software (DIFF\&GEO), and Land Application Tools (LAT) are modular software packages useful to process synthetic aperture radar (SAR) images. Data of both spaceborne and airborne sensors including ERS-1/2, JERS-1, SIRC, SEASAT, RADARSAT StripMap mode, and the single-pass Dornier DOSAR interferometer have been successfully processed interferometrically. State of the art algorithms have been implemented to achieve accurate processing of the data while permitting timely processing of large data sets. Recent projects completed with the software include generation of a continental scale mosaic of Siberia consisting of more 700 JERS scenes in the frame of the SIBERIA project and the generation of subsidence maps for Bologna, Abano, and Mexico City. User-friendly display tools and full documentation in HTML language complements the software. Both binary and source code licenses are available.},
     language = {en},
     author = {Werner, Charles and Wegmüller, Urs and Strozzi, Tazio and Wiesmann, Andreas},
@@ -122,7 +122,6 @@ The overall aim is thus to reduce the entire process of image processing to very
     language = {en},
     urldate = {2021-11-29},
     author = {Truckenbrodt, John and Cremer, Felix and Baris, Ismail and Glaser, Felix and Eberle, Jonas},
-    month = may,
     year = {2019},
     file = {Full Text PDF:/home/wodan/snap/zotero-snap/common/Zotero/storage/K4UHSLZH/Truckenbrodt et al. - 2019 - pyroSAR - a Python Framework for Large-Scale SAR S.pdf:application/pdf;Snapshot:/home/wodan/snap/zotero-snap/common/Zotero/storage/3YJKTKL6/133267.html:text/html},
 }
