### 2021-11-29 #3337 @McWhity
> Apologies this has taken so long.
>
> I'm happy to recommend **accept with minor revisions** if the paper is updated to clarify who it is aimed at and how it differs from other packages.

@danclewley: Thank you for the review. I updated the paper (https://github.com/multiply-org/sar-pre-processing/blob/master/docs/paper.pdf).

Main changes are:
- last paragraph of method section was rewritten:
> After the pre-processing the resulting radiometrically and geometrically corrected images are stored for further usage within a NetCDF4 stack file. The processing workflow was developed and optimized to use a Sentinel-1 time series of pre-processed sigma nought backscatter values to retrieve biophysical land surface parameters by the use of radiative transfer models. The sigma nought backscatter values provided by the default workflow of SenSARP might be used in other applications like flood risk analysis, monitoring land cover changes or monitoring global food security but it has to be mentioned that different applications have different demands and therefore, slight adjustments of the default workflow might be required. In the future, many more new products and operational third party services based on consistent Sentinel-1 time series might be developed._

- new Section regarding other python packages was added:

> **Other available python software packages using ESA's SNAP software to pre-process SAR data**
The ESA's SNAP toolbox has been written in Java. For python users the developers provide a python interface called Snappy. However, the Snappy interface is lacking in terms of installation, processing performance and usability. Hence, the remote sensing community developed different wrappers (e.g. SenSARP, snapista or pyroSAR) to use SNAP processing functionalities by utilizing the SNAP Graph Processing Tool (GPT).

> **snapista**
Snapista (https://snap-contrib.github.io/snapista/index.html) targets mainly experts remote sensing users with python programming skills. It provides access to the processing operators of all toolboxes (e.g. Sentinel-1, Sentinel-2 or Sentinel-3) within SNAP. Expert users can generate processing graphs and execute there generated graphs in a pure Pythonic way. A guideline which processing steps are needed for different applications or which processing steps can or have to be combined are not provided yet. As guidelines how to process different satellite data for different applications is not an easy task to do it exceeds the goal of snapista as a python wrapper for the SNAP software. Summarizing, snapista provides access to all SNAP toolboxes (not just to Sentinel-1 Toolbox) via python. But as it provides no default processing chains, snapista will be primarily usable by expert remote sensing users. The advantage of snapista is the accessibility of processing operators for SAR and optical data.

> **pyroSAR**
PyroSAR (https://pyrosar.readthedocs.io/en/latest/index.html) is a python library which provides a python wrapper to SAR pre-processing software SNAP and GAMMA [@wegnuller_sentinel-1_2016; @werner_gamma_2000]. The library provides utilities to read and store metadata information of downloaded satellite data within a database. Furthermore, pyroSAR provides access to processing operators of SNAP and GAMMA. A default workflow with different user options is provided to process single or time-series Sentinel-1 images. After executing the default processing workflow radiometric and geometric corrected gamma nought backscatter data are provided in Geotiff format [@truckenbrodt_pyrosar_2019]. The processed images can also be stored within an Open Data Cube. For expert users which might want to use a different processing workflow pyroSAR provides an option to create SNAP xml-workflows and execute them with the GPT. Summarizing, pyroSAR provides a similar push-button option to process Sentinel-1 data with a slightly different default workflow (pyroSAR: no temporal speckle filter, gamma nought backscatter output in Geotiff format) than SenSARP (SenSARP: temporal speckle filter, sigma nought backscatter output in netCDF format). PyroSAR, as a more complex library than SenSARP, provides on the one hand more changeable parameters within the processing workflow but on the other hand the usability for non-expert users might be narrowed compared to SenSARP. An advantage of SenSARP, especially for non-expert users, might be the provision of background information (theory/purpose) of the different pre-processing steps within the documentation.

I hope these changes satisfying your request
