### 2021-11-29 multiply-org/sar-pre-processing#13 @McWhity
I think the new version of sentinesat and the provision of an alternative (Alaska Satellite Facility) source has solved this issue. If there are no objections, I would like to close this issue.
