### 2021-11-29 multiply-org/sar-pre-processing#12 @arthur-e
@McWhity I have not yet had a chance to evaluate these changes to the code. I have no objections to closing it; if future users experience a similar problem, they can re-open this issue.
