### 2021-12-05 #3337 @danclewley
Updates look good to me and address my comments. Happy to recommend accepting.
