### 2021-12-05 #3337 @pdebuyl
Hi @danclewley thanks for the feedback! Can you check all relevant checkboxes? This is part of the JOSS process and despite your message I don't think it would be right for me to check them :-)
