### 2022-01-12 #3337 @pdebuyl
I have a minor comment for the installation: when using pip, it seems that you need to pick a pygdal version that is of same "point release" (in `x.y.z` only `x` and `y` need to match) so that the recommended methods can lead to incomprehension from users. Of course, nowadays using pip for scientific packages is more of an expert choice unfortunately.
