### 2022-01-15 #3337 @McWhity
The installation via virtualenv and pip was added due to a comment by arthur-e.

> - I would prefer it if there were some way to install the software without using conda, which can be a pain to work with (and very slow). In this case, conda took several minutes to install the repository (spent a lot of time on "Solving environment"). Ideally, the setup.py script would list the package dependencies so that the repository (and its dependencies) could be installed with either distutils or pip. This can still be done within a conda virtual environment or, better yet, a virtualenv environment.

The installation of a working gdal version can be tricky. If using pip, `x.y.z` of the pygdal version (`x.y.z.buildnumber)` need to be the same as the previously installed `gdal-bin x.y.z` package. Does this clarify your concerns?

Beside the installation via pip and virtualenv a second installation guide by using Anaconda/Miniconda and conda is given in the documentary.
