### 2022-01-15 #3337 @arthur-e
@McWhity Yes, that's great. Giving users options is always helpful. Power users will be able to install the GDAL Python bindings ahead of time and then use whatever package manager they want for the rest. Less experienced users will have to wait on Conda.
