### 2022-01-16 #3337 @pdebuyl
Hi @McWhity

- I filed a pull request for minor edits to the article. I think that the changes do not modify the intent of your text, let me know if this is the case.
- Can you archive the software on zenodo and provide here the version information of that archive and zenodo's DOI ?
- The last reference has no URL or DOI, could provide a stronger identification?

Once those items are checked, I'll forward the article for acceptance.

Regarding the installation: I admit that the specific annoyance of GDAL install goes beyond the scope of SenSARP, thanks for the clarification. I could install with the debian-provided GDAL in the end, which is my default choice :-)
