### 2022-01-17 #3337 @pdebuyl
I must have missed a clic. Fortunately, the patch was still saved by GitHub's web interface, I opened the PR.
