+++ docs/paper.bib
@@ -79,14 +79,13 @@
 }
 
 @inproceedings{werner_gamma_2000,
-    address = {Gothenburg Sweden},
-    title = {{GAMMA} {SAR} {and} {Interferometric} {Processing} {Software}},
-    abstract = {The GAMMA Modular SAR Processor (MSP), Interferometric SAR Processor (ISP), Differential Interferometry and Geocoding Software (DIFF\&GEO), and Land Application Tools (LAT) are modular software packages useful to process synthetic aperture radar (SAR) images. Data of both spaceborne and airborne sensors including ERS-1/2, JERS-1, SIRC, SEASAT, RADARSAT StripMap mode, and the single-pass Dornier DOSAR interferometer have been successfully processed interferometrically. State of the art algorithms have been implemented to achieve accurate processing of the data while permitting timely processing of large data sets. Recent projects completed with the software include generation of a continental scale mosaic of Siberia consisting of more 700 JERS scenes in the frame of the SIBERIA project and the generation of subsidence maps for Bologna, Abano, and Mexico City. User-friendly display tools and full documentation in HTML language complements the software. Both binary and source code licenses are available.},
-    language = {en},
-    author = {Werner, Charles and Wegmüller, Urs and Strozzi, Tazio and Wiesmann, Andreas},
-    year = {2000},
-    pages = {9},
-    file = {Werner et al. - 2000 - GAMMA SAR AND INTERFEROMETRIC PROCESSING SOFTWARE.pdf:/home/wodan/snap/zotero-snap/common/Zotero/storage/KGZLVNET/Werner et al. - 2000 - GAMMA SAR AND INTERFEROMETRIC PROCESSING SOFTWARE.pdf:application/pdf},
+  title={Gamma SAR and interferometric processing software},
+  author={Werner, Charles and Wegm{\"u}ller, Urs and Strozzi, Tazio and Wiesmann, Andreas},
+  booktitle={Proceedings of the ers-envisat symposium, Gothenburg, Sweden},
+  url = {https://www.gamma-rs.ch/uploads/media/2000-1_GAMMA_Software.pdf},
+  volume={1620},
+  pages={9},
+  year={2000},
 }
 
 @article{wegnuller_sentinel-1_2016,
