### 2022-01-18 #3337 @McWhity
- PR is merged. Thank you for the revision
- last reference was revised (URL was added)
- changelog was updated and new tag v0.2 was created

As the Github repository is within a Github organization I need permission from the organization owner to grant zenodo access. As soon as the permission is granted, I will archive SenSARP on zenodo and provide the required information. This shouldn't take long.
