### 2022-01-25 #3337 @McWhity
@pdebuyl: SenSARP is now archived on zenodo. The DOI is [10.5281/zenodo.5903483](https://zenodo.org/record/5903484) (Version v0.2).
