### 2022-01-25 #3337 @pdebuyl
Note to the EiC: I accepted to perform the functionality check in place of the second reviewer for availability reasons.
