### 2022-01-29 #3337 @arfon
[@arthur-e](https://github.com/arthur-e), [@danclewley](https://github.com/danclewley) – many thanks for your reviews here and to @pdebuyl for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
